renamed corgiPhoto to corgiPhotos, deleted Photo.jsx component due to similar components in corgiPhoto folder, moved SwitchCorgiPhoto component in corgiPhoto folder
