--- src/component/utils/Photo.jsx
@@ -1,103 +1,0 @@
-import React from 'react';
-
-import { CorgiType } from '../../types/CorgiTypes';
-
-const CommonPropTypes = { color: CorgiType.color };
-
-export const Common = ({ color }) => (
-  <div>
-    <svg
-      width='72px'
-      height='48px'
-      viewBox='0 0 72 48'
-      version='1.1'
-      xmlns='http://www.w3.org/2000/svg'
-      xlink='http://www.w3.org/1999/xlink'
-    >
-      <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-        <g id='result' transform='translate(-121.000000, -1034.000000)' fill={color}>
-          <path
-            d='M139.074948,1079.0508 C139.088948,1079.01813 139.103615,1078.98547 139.118282,1078.9528 L139.120282,1078.9488 C139.135615,1078.91547 139.150948,1078.8808 139.167615,1078.8468 C139.167615,1078.84613 139.167615,1078.84547 139.168282,1078.8448 C139.183615,1078.81013 139.200282,1078.7748 139.217615,1078.73813 C139.344948,1078.4648 139.494948,1078.1608 139.663615,1077.84547 C139.637615,1078.10213 139.588948,1078.35813 139.519615,1078.60813 C139.446948,1078.8708 139.352948,1079.1348 139.241615,1079.38013 C139.140948,1079.3708 139.038948,1079.3668 138.937615,1079.36547 C138.978282,1079.26947 139.024282,1079.1648 139.074948,1079.0508 M192.956948,1045.00013 C192.833615,1043.45547 192.507615,1041.92547 191.969615,1040.47347 C191.703615,1039.74613 191.390282,1039.0368 191.033615,1038.34947 C190.857615,1038.00547 190.666282,1037.66813 190.464948,1037.3368 C190.262282,1037.0048 190.054282,1036.6828 189.813615,1036.36547 C189.783615,1036.3248 189.735615,1036.29813 189.681615,1036.2968 C189.586948,1036.29413 189.506948,1036.36947 189.504948,1036.46413 L189.504948,1036.47147 C189.502948,1036.5288 189.496948,1036.6028 189.489615,1036.6708 C189.483615,1036.73947 189.472948,1036.81013 189.464282,1036.88013 C189.445615,1037.02013 189.420948,1037.1608 189.394948,1037.3008 C189.342282,1037.58147 189.280282,1037.86147 189.212948,1038.1408 C189.143615,1038.41947 189.072282,1038.69813 188.992282,1038.9748 C188.916948,1039.2528 188.829615,1039.52747 188.745615,1039.80347 C188.401615,1040.90413 188.004948,1041.99013 187.586282,1043.0668 L187.266948,1043.87347 L186.935615,1044.67547 C186.720282,1045.21147 186.485615,1045.74013 186.258282,1046.2728 C185.808948,1047.33147 185.242948,1048.3348 184.526948,1049.2348 C183.818282,1050.1388 182.996948,1050.95347 182.102282,1051.67747 C181.648282,1052.0328 181.187615,1052.37947 180.702282,1052.69147 C180.224282,1053.0128 179.726282,1053.30613 179.221615,1053.58613 C178.714948,1053.86347 178.196282,1054.1188 177.671615,1054.36013 C177.144282,1054.59613 176.609615,1054.8168 176.066948,1055.01947 C175.714948,1055.15147 175.358282,1055.27347 175.000282,1055.38947 C174.933615,1055.3288 174.866948,1055.27013 174.800282,1055.21147 C174.767615,1055.18213 174.734948,1055.15213 174.701615,1055.1228 C174.637615,1055.06747 174.572948,1055.01413 174.508282,1054.95947 C174.468948,1054.92613 174.428948,1054.89213 174.388948,1054.85947 C174.326948,1054.80813 174.263615,1054.7588 174.200948,1054.7088 C174.158282,1054.67547 174.115615,1054.64013 174.072282,1054.60613 C174.010282,1054.5588 173.948948,1054.51213 173.886948,1054.46547 C173.841615,1054.43147 173.796948,1054.39813 173.751615,1054.36413 C173.690282,1054.3188 173.628948,1054.27613 173.566948,1054.23147 C173.520282,1054.1988 173.474282,1054.16547 173.426948,1054.1328 C173.366282,1054.09013 173.304282,1054.0488 173.242282,1054.00813 C173.194948,1053.97547 173.146948,1053.9428 173.098948,1053.91147 C173.037615,1053.8708 172.976282,1053.83213 172.914282,1053.7928 C172.865615,1053.76213 172.816948,1053.7308 172.768282,1053.7008 C172.706282,1053.6628 172.644282,1053.62547 172.582948,1053.5888 C172.533615,1053.5588 172.483615,1053.5288 172.433615,1053.49947 C172.371615,1053.46347 172.310282,1053.42813 172.248282,1053.39347 C172.197615,1053.3648 172.146948,1053.33613 172.096282,1053.30813 C172.034282,1053.27413 171.972282,1053.2408 171.909615,1053.2068 C171.858948,1053.18013 171.808282,1053.1528 171.756948,1053.1268 C171.694948,1053.09413 171.632282,1053.0628 171.568948,1053.03147 C171.517615,1053.00547 171.466948,1052.97947 171.415615,1052.95413 C171.352282,1052.92413 171.288948,1052.89413 171.226948,1052.86413 C171.174282,1052.84013 171.122948,1052.8148 171.070948,1052.7908 C171.007615,1052.76213 170.944948,1052.73413 170.881615,1052.70547 C170.829615,1052.6828 170.777615,1052.65947 170.724948,1052.6368 L170.534948,1052.5568 C170.482282,1052.53547 170.429615,1052.5128 170.376948,1052.49147 C170.313615,1052.46613 170.250282,1052.4408 170.186948,1052.41613 C170.134282,1052.39547 170.081615,1052.3748 170.028948,1052.3548 C169.964948,1052.3308 169.900948,1052.30747 169.837615,1052.28413 C169.783615,1052.26413 169.730948,1052.24547 169.677615,1052.2268 C169.614282,1052.20347 169.550282,1052.18147 169.486282,1052.15947 C169.432948,1052.14213 169.380282,1052.12347 169.326948,1052.10613 C169.262948,1052.0848 169.198948,1052.0648 169.134948,1052.04347 C169.081615,1052.0268 169.028282,1052.01013 168.974948,1051.99347 C168.910948,1051.97413 168.846282,1051.95413 168.782282,1051.93547 C168.728948,1051.92013 168.675615,1051.90413 168.622282,1051.88947 C168.558282,1051.8708 168.494282,1051.8528 168.429615,1051.83547 C168.376282,1051.82013 168.322948,1051.80613 168.269615,1051.79213 C168.204948,1051.77547 168.140948,1051.7588 168.076948,1051.74213 C168.023615,1051.7288 167.970282,1051.71547 167.917615,1051.7028 C167.852282,1051.6868 167.787615,1051.67147 167.723615,1051.65613 C167.670948,1051.64413 167.617615,1051.63147 167.564282,1051.62013 L167.370948,1051.5768 C167.318282,1051.56547 167.264948,1051.55547 167.212282,1051.54413 C167.147615,1051.5308 167.083615,1051.51747 167.018948,1051.50547 C166.966282,1051.4948 166.913615,1051.4848 166.860948,1051.47547 C166.796282,1051.4628 166.731615,1051.45147 166.667615,1051.43947 C166.615615,1051.43013 166.562282,1051.42213 166.510282,1051.4128 C166.446282,1051.40147 166.381615,1051.3908 166.316948,1051.3808 C166.264948,1051.37213 166.212948,1051.3648 166.161615,1051.3568 C166.096948,1051.3468 166.032282,1051.33747 165.967615,1051.32813 C165.916282,1051.32013 165.864948,1051.31413 165.812948,1051.3068 C165.748948,1051.29813 165.684282,1051.2888 165.620282,1051.28147 C165.568948,1051.2748 165.518282,1051.2688 165.466948,1051.2628 C165.402282,1051.25547 165.338282,1051.24747 165.273615,1051.24013 C165.224282,1051.23413 165.174282,1051.23013 165.124282,1051.2248 C165.059615,1051.21813 164.994282,1051.2108 164.928948,1051.20413 C164.880948,1051.20013 164.833615,1051.19613 164.784948,1051.19213 C164.718948,1051.18613 164.652948,1051.18013 164.586948,1051.1748 L164.448282,1051.1648 C164.380948,1051.15947 164.313615,1051.15413 164.247615,1051.14947 C164.202282,1051.14747 164.158282,1051.1448 164.114282,1051.14213 C164.046282,1051.13813 163.977615,1051.13347 163.910282,1051.13013 C163.869615,1051.12813 163.830282,1051.1268 163.789615,1051.1248 C163.718282,1051.12147 163.646282,1051.11813 163.575615,1051.11547 C163.548948,1051.1148 163.522282,1051.11413 163.496282,1051.1128 C163.220948,1051.10347 162.948948,1051.09813 162.680282,1051.09747 C162.652948,1051.09747 162.625615,1051.0968 162.598948,1051.0968 C162.531615,1051.0968 162.464948,1051.09747 162.397615,1051.09813 C162.357615,1051.0988 162.316948,1051.0988 162.277615,1051.09947 C162.215615,1051.10013 162.154282,1051.10147 162.092948,1051.1028 C162.048948,1051.10347 162.004282,1051.10413 161.959615,1051.10547 C161.901615,1051.1068 161.844282,1051.1088 161.786948,1051.1108 C161.739615,1051.11213 161.692948,1051.11347 161.645615,1051.1148 C161.588948,1051.11747 161.533615,1051.11947 161.478282,1051.12147 C161.430948,1051.12347 161.384282,1051.12547 161.337615,1051.12747 C161.282282,1051.13013 161.227615,1051.13347 161.172282,1051.13613 C161.125615,1051.1388 161.078948,1051.14147 161.032948,1051.14413 C160.978282,1051.14747 160.925615,1051.15013 160.872282,1051.15347 C160.825615,1051.1568 160.779615,1051.15947 160.732948,1051.1628 L160.576948,1051.1748 C160.530948,1051.17747 160.484282,1051.1808 160.438282,1051.1848 C160.387615,1051.1888 160.337615,1051.19347 160.286948,1051.19747 C160.240948,1051.20147 160.194948,1051.2048 160.148948,1051.2088 C160.099615,1051.21347 160.050948,1051.21813 160.002282,1051.2228 C159.956282,1051.22747 159.910948,1051.23147 159.865615,1051.23547 C159.816948,1051.2408 159.769615,1051.24613 159.721615,1051.2508 C159.676948,1051.25547 159.632282,1051.26013 159.588282,1051.2648 C159.540282,1051.27013 159.494282,1051.27547 159.447615,1051.2808 C159.402948,1051.28613 159.359615,1051.2908 159.314948,1051.29613 C159.270282,1051.30147 159.224948,1051.3068 159.179615,1051.3128 C159.136282,1051.31747 159.092948,1051.3228 159.049615,1051.3288 C159.005615,1051.33413 158.960948,1051.34013 158.917615,1051.34547 C158.874948,1051.35147 158.832948,1051.3568 158.790948,1051.3628 C158.747615,1051.3688 158.704282,1051.3748 158.661615,1051.3808 C158.620282,1051.3868 158.578948,1051.3928 158.538282,1051.39813 C158.496282,1051.40413 158.454948,1051.4108 158.413615,1051.4168 C158.372948,1051.42347 158.332948,1051.4288 158.292948,1051.4348 C158.252282,1051.44147 158.211615,1051.44813 158.171615,1051.45347 C158.132282,1051.46013 158.092948,1051.4668 158.054948,1051.4728 C158.014948,1051.47947 157.976282,1051.48547 157.937615,1051.49213 C157.898948,1051.4988 157.860948,1051.50547 157.822948,1051.5108 C157.784948,1051.51747 157.748282,1051.52413 157.711615,1051.5308 C157.672948,1051.5368 157.635615,1051.54347 157.598282,1051.5508 C157.562282,1051.5568 157.527615,1051.56347 157.492948,1051.56947 C157.455615,1051.57613 157.419615,1051.5828 157.382948,1051.59013 C157.348282,1051.59613 157.313615,1051.6028 157.280282,1051.6088 C157.245615,1051.61613 157.210948,1051.62213 157.176282,1051.6288 L157.076948,1051.64813 C157.042948,1051.6548 157.009615,1051.66213 156.976948,1051.6688 C156.944948,1051.67547 156.913615,1051.6808 156.882282,1051.68747 C156.849615,1051.69413 156.817615,1051.70147 156.785615,1051.70747 C156.755615,1051.71347 156.726282,1051.72013 156.696948,1051.7268 C156.664948,1051.7328 156.634282,1051.73947 156.602948,1051.74613 C156.574948,1051.7528 156.546948,1051.7588 156.518948,1051.76413 C156.488948,1051.77147 156.460282,1051.77747 156.430948,1051.7848 C156.412282,1051.78813 156.392948,1051.79213 156.373615,1051.7968 C155.396948,1052.0168 154.399615,1052.13147 153.398282,1052.13147 L150.477615,1052.13147 C147.060948,1050.87413 144.922948,1052.28147 144.922948,1052.28147 C143.187615,1049.9668 140.982282,1047.1848 138.698282,1044.8248 C138.632282,1044.75747 138.566948,1044.6888 138.501615,1044.62147 C139.114948,1043.22013 138.644282,1041.1108 138.644282,1041.1108 C139.160948,1039.62213 139.616282,1035.18613 137.489615,1034.1528 C135.362948,1033.12013 134.207615,1037.6168 133.582282,1038.80147 C132.956282,1039.9868 132.394282,1039.8048 132.394282,1039.8048 C132.570948,1040.13147 132.496282,1040.45747 132.397615,1040.67947 C131.870282,1040.6868 131.273615,1040.72613 130.614282,1040.74813 C130.067615,1040.76613 129.478282,1040.77213 128.847615,1040.7388 C128.575615,1040.7248 128.296282,1040.70347 128.008282,1040.67213 C127.980282,1040.64613 127.950282,1040.6168 127.916948,1040.58613 C127.852282,1040.52613 127.778282,1040.4568 127.696948,1040.3808 C127.685615,1040.37013 127.672948,1040.35947 127.662282,1040.3488 C127.632948,1040.32147 127.603615,1040.29347 127.572948,1040.26413 C127.557615,1040.24947 127.540282,1040.23347 127.524948,1040.21813 C127.496948,1040.19213 127.469615,1040.16547 127.440948,1040.13747 C127.422282,1040.12013 127.403615,1040.10147 127.385615,1040.08347 C127.357615,1040.05613 127.329615,1040.02813 127.300948,1039.99947 C127.281615,1039.9808 127.262282,1039.9608 127.242948,1039.94147 C127.213615,1039.91147 127.184282,1039.8808 127.154282,1039.85013 L127.097615,1039.79147 C127.064948,1039.75747 127.032282,1039.72213 126.999615,1039.68613 C126.982948,1039.66947 126.967615,1039.65213 126.950948,1039.6348 C126.902282,1039.58147 126.853615,1039.5268 126.805615,1039.4708 C126.803615,1039.4688 126.802282,1039.4668 126.800282,1039.4648 C126.754282,1039.41147 126.708282,1039.3568 126.662948,1039.30147 C126.646948,1039.2828 126.632948,1039.26347 126.617615,1039.2448 C126.586282,1039.20547 126.554948,1039.16613 126.524282,1039.12547 C126.506282,1039.1028 126.489615,1039.08013 126.472282,1039.0568 C126.444948,1039.01947 126.417615,1038.98213 126.390948,1038.9448 C126.372948,1038.92013 126.356282,1038.89547 126.338948,1038.87013 C126.312948,1038.83347 126.288282,1038.79547 126.264282,1038.75813 C126.248282,1038.73213 126.231615,1038.7068 126.216282,1038.68147 C126.192282,1038.64347 126.169615,1038.60413 126.146282,1038.5648 C126.132282,1038.54013 126.117615,1038.51547 126.104282,1038.4908 C126.080948,1038.44813 126.059615,1038.4048 126.037615,1038.36213 C126.026948,1038.3408 126.015615,1038.3188 126.005615,1038.29747 C125.975615,1038.2328 125.946948,1038.16813 125.921615,1038.1028 C125.366282,1036.67547 123.603615,1033.9708 122.175615,1034.57813 C120.747615,1035.18613 121.020948,1038.83213 121.020948,1038.83213 C121.020948,1038.83213 121.222282,1040.07813 121,1040.38147 C121,1040.38147 121.385615,1040.74613 121.537615,1041.50613 C121.669615,1042.16547 121.247615,1044.33947 122.320948,1045.25613 C122.071615,1045.7748 121.936948,1046.12013 121.936948,1046.12013 C121.936948,1046.12013 122.266948,1047.3908 121.172948,1047.7148 C121.172948,1047.7148 122.590948,1047.7148 122.388282,1048.5648 C122.185615,1049.41613 121.780948,1051.2568 121.982948,1052.28147 C121.982948,1052.28147 121.989615,1052.27213 122.000948,1052.25813 C122.012282,1052.24213 122.030282,1052.22013 122.052282,1052.19413 C122.064948,1052.17947 122.078948,1052.16347 122.094282,1052.14747 C122.112282,1052.12747 122.131615,1052.10813 122.152948,1052.08813 C122.164948,1052.07613 122.177615,1052.0648 122.190948,1052.05347 C122.243615,1052.0068 122.303615,1051.96213 122.366282,1051.93147 C122.398948,1051.91547 122.433615,1051.9028 122.468282,1051.89613 C122.487615,1051.89147 122.506948,1051.8888 122.526948,1051.88813 C122.546948,1051.88747 122.566948,1051.89013 122.587615,1051.89413 C122.593615,1051.89547 122.600948,1051.89613 122.607615,1051.89813 C122.630282,1051.90413 122.652948,1051.91547 122.674948,1051.9288 C122.678948,1051.9308 122.682948,1051.9328 122.686282,1051.93547 C122.691615,1051.9388 122.696282,1051.9428 122.700282,1051.9468 C122.708948,1051.9528 122.716948,1051.95813 122.724948,1051.96547 C122.726948,1051.9668 122.727615,1051.9688 122.729615,1051.9708 C122.753615,1051.9928 122.776282,1052.01947 122.798282,1052.0528 C122.816948,1052.0808 122.834948,1052.11413 122.852282,1052.1508 C122.856948,1052.16013 122.861615,1052.1688 122.865615,1052.17813 C122.866948,1052.18213 122.868948,1052.18747 122.870948,1052.19213 C122.900282,1052.2628 122.927615,1052.34747 122.951615,1052.44947 C122.963615,1052.49813 122.971615,1052.54747 122.979615,1052.5968 C122.984948,1052.6328 122.990282,1052.6688 122.994948,1052.70413 C123.014948,1052.8788 123.018282,1053.0548 123.003615,1053.23013 C122.960948,1053.7268 122.885615,1054.88613 122.935615,1056.42413 L122.935615,1056.4248 C122.936282,1056.43147 122.936282,1056.43947 122.936948,1056.44747 C122.938948,1056.52013 122.942282,1056.59347 122.944948,1056.6688 L122.948948,1056.7568 C122.950948,1056.8108 122.952948,1056.8648 122.955615,1056.91947 C122.957615,1056.95347 122.959615,1056.98747 122.960948,1057.02213 C122.964282,1057.07347 122.966948,1057.1248 122.969615,1057.1768 C122.971615,1057.2128 122.974282,1057.24947 122.976282,1057.28547 C122.978948,1057.33747 122.982282,1057.38947 122.985615,1057.44213 C122.988282,1057.4788 122.990948,1057.51547 122.992948,1057.5528 C122.996948,1057.60613 123.000948,1057.66013 123.004948,1057.71413 C123.006948,1057.7508 123.010282,1057.7868 123.012948,1057.82347 C123.017615,1057.8828 123.022948,1057.94347 123.027615,1058.00347 C123.030282,1058.03547 123.032948,1058.06547 123.034948,1058.09747 C123.043615,1058.19013 123.052282,1058.2828 123.060948,1058.37613 C123.060948,1058.3788 123.060948,1058.38213 123.061615,1058.3848 C123.070282,1058.47613 123.079615,1058.56747 123.088948,1058.65947 C123.092282,1058.6908 123.096282,1058.7228 123.099615,1058.75413 C123.106948,1058.81813 123.114282,1058.8828 123.120948,1058.94813 C123.125615,1058.9848 123.130282,1059.0228 123.134948,1059.0608 C123.142282,1059.12013 123.148948,1059.18013 123.156948,1059.2408 C123.161615,1059.2808 123.166948,1059.32013 123.172282,1059.36013 C123.179615,1059.41947 123.187615,1059.47947 123.196282,1059.5388 C123.200948,1059.57947 123.206948,1059.62013 123.212282,1059.66013 C123.220948,1059.72147 123.229615,1059.78347 123.238948,1059.84413 C123.244948,1059.88347 123.250948,1059.9228 123.256282,1059.96213 C123.266948,1060.02947 123.277615,1060.0968 123.288282,1060.1648 C123.293615,1060.1988 123.298948,1060.23147 123.304282,1060.26547 C123.320282,1060.36747 123.338282,1060.46947 123.355615,1060.57213 C123.355615,1060.57213 124.252282,1067.40347 127.106948,1070.1188 C127.106948,1070.1188 126.925615,1069.65613 126.978282,1069.34413 C131.507615,1075.41147 134.722948,1075.68413 134.722948,1075.68413 L136.100948,1078.25147 L136.052282,1078.49547 C135.956948,1078.51213 135.856282,1078.5328 135.751615,1078.5588 C135.490948,1078.62147 135.208282,1078.71547 134.943615,1078.85347 C134.890948,1078.8808 134.838948,1078.91013 134.786948,1078.94147 C134.696282,1078.99747 134.608282,1079.06013 134.524948,1079.12813 C134.426282,1079.2088 134.334948,1079.29813 134.253615,1079.39813 C134.092282,1079.59413 133.968948,1079.82947 133.904282,1080.11213 C133.862948,1080.29147 133.956282,1080.4648 134.109615,1080.54013 C134.160282,1080.56547 134.218282,1080.58013 134.279615,1080.58013 L136.833615,1080.58013 C136.791615,1080.67547 136.755615,1080.77147 136.726948,1080.87013 C136.707615,1080.93547 136.691615,1081.00213 136.677615,1081.0688 L136.668282,1081.11947 C136.665615,1081.13547 136.662282,1081.1528 136.659615,1081.17613 C136.653615,1081.21947 136.654948,1081.26413 136.660282,1081.30813 C136.682282,1081.4828 136.800282,1081.64147 136.962282,1081.71347 C137.002948,1081.73147 137.044948,1081.74547 137.088948,1081.75213 C137.130282,1081.76013 137.180948,1081.75947 137.208282,1081.75947 L137.792282,1081.7588 L139.349615,1081.75413 L142.464948,1081.74547 L142.659615,1081.7448 L142.756948,1081.74413 C142.790948,1081.74413 142.816948,1081.74547 142.864282,1081.7428 C143.032948,1081.73013 143.194282,1081.62747 143.272282,1081.47213 C143.290948,1081.43547 143.306948,1081.3868 143.312948,1081.35747 C143.321615,1081.32413 143.330948,1081.2908 143.337615,1081.25813 C143.352282,1081.1908 143.364282,1081.12413 143.372948,1081.05747 C143.392282,1080.92347 143.400948,1080.7888 143.404282,1080.6548 C143.408282,1080.13947 143.332948,1079.63813 143.204948,1079.1428 C143.716282,1078.62547 144.106282,1078.0028 144.391615,1077.34413 C144.546282,1076.9868 144.672282,1076.6188 144.776282,1076.24413 C145.059615,1077.09213 144.690948,1077.93747 144.690948,1077.93747 C146.686948,1077.6048 148.124282,1076.25347 148.124282,1076.25347 C148.930948,1076.53813 149.314282,1076.53947 149.314282,1076.53947 C149.932282,1076.15947 150.477615,1076.25347 150.477615,1076.25347 L153.338282,1076.25347 C154.170282,1076.25347 153.615615,1076.80813 153.615615,1076.80813 C155.468948,1076.80813 156.906282,1075.2348 156.906282,1075.2348 L156.666948,1075.96947 C159.565615,1074.78147 159.803615,1074.87613 159.803615,1074.87613 C159.751615,1075.12547 159.613615,1075.2668 159.613615,1075.2668 C159.714282,1075.22547 159.811615,1075.18213 159.906948,1075.1388 C159.935615,1075.12547 159.962282,1075.11213 159.990948,1075.0988 C160.056282,1075.06813 160.120948,1075.03813 160.183615,1075.00747 C160.215615,1074.99147 160.245615,1074.97613 160.276948,1074.95947 C160.333615,1074.9308 160.390282,1074.9028 160.444282,1074.87347 C160.476948,1074.8568 160.506282,1074.84013 160.537615,1074.82213 C160.588948,1074.7948 160.640282,1074.7668 160.688948,1074.7388 C160.720282,1074.72147 160.748948,1074.70347 160.779615,1074.68613 C160.826282,1074.65813 160.873615,1074.6308 160.918948,1074.60347 C160.948948,1074.5848 160.977615,1074.56613 161.006948,1074.54747 C161.048948,1074.52147 161.091615,1074.4948 161.132282,1074.46813 C161.162948,1074.4488 161.190948,1074.4288 161.220282,1074.40947 C161.256948,1074.38413 161.295615,1074.35947 161.331615,1074.33413 C161.361615,1074.31347 161.389615,1074.2928 161.418282,1074.27213 C161.450948,1074.2488 161.484282,1074.22547 161.515615,1074.20213 C161.545615,1074.17947 161.572948,1074.15813 161.601615,1074.1368 C161.629615,1074.11547 161.658948,1074.0928 161.686282,1074.07147 C161.715615,1074.0488 161.742948,1074.0268 161.770948,1074.00413 C161.794948,1073.98413 161.819615,1073.9648 161.842948,1073.94547 C161.872282,1073.9208 161.900282,1073.8968 161.928282,1073.87213 C161.947615,1073.85613 161.967615,1073.83947 161.985615,1073.8228 C162.015615,1073.7968 162.043615,1073.77147 162.071615,1073.74547 C162.086282,1073.73213 162.101615,1073.7188 162.116282,1073.70547 C162.144948,1073.67747 162.172948,1073.6508 162.200948,1073.62413 C162.211615,1073.61413 162.222948,1073.60347 162.232948,1073.59347 C162.262282,1073.5648 162.290282,1073.5368 162.317615,1073.5088 C162.324282,1073.50213 162.331615,1073.4948 162.337615,1073.48813 C162.366948,1073.45747 162.394282,1073.42813 162.420948,1073.40013 C162.423615,1073.3968 162.426948,1073.39347 162.430282,1073.39013 C162.436282,1073.38347 162.442948,1073.37613 162.448948,1073.37013 C162.756282,1073.7168 163.084948,1074.04613 163.442282,1074.34613 L163.692282,1074.55213 C163.778282,1074.61813 163.868282,1074.6768 163.956282,1074.74013 C164.044948,1074.8028 164.132948,1074.86613 164.228948,1074.91747 C164.323615,1074.97147 164.415615,1075.03013 164.512282,1075.07947 L164.809615,1075.21947 C164.913615,1075.25747 165.014948,1075.2988 165.120282,1075.33147 C165.228282,1075.36013 165.334282,1075.3928 165.446948,1075.40813 L165.530282,1075.42213 L165.616282,1075.4288 C165.634282,1075.43013 165.652282,1075.43147 165.671615,1075.43213 C165.717615,1075.80947 165.752948,1076.19013 165.778948,1076.56947 C165.806948,1076.9868 165.822948,1077.40547 165.818948,1077.82213 C165.813615,1078.23813 165.791615,1078.6548 165.719615,1079.05747 C165.683615,1079.25813 165.634282,1079.45547 165.556948,1079.6328 C165.484282,1079.79813 165.377615,1079.93813 165.248282,1080.00147 C164.943615,1079.94413 164.644948,1079.92213 164.338948,1079.93213 C164.019615,1079.94547 163.694282,1080.00013 163.388948,1080.12947 C163.236282,1080.1948 163.089615,1080.28013 162.958948,1080.38747 C162.828948,1080.49547 162.716282,1080.62547 162.628282,1080.76947 C162.540948,1080.91413 162.473615,1081.07013 162.433615,1081.2308 C162.427615,1081.25147 162.422282,1081.2688 162.416948,1081.29213 C162.412948,1081.3168 162.406948,1081.34147 162.405615,1081.3668 C162.401615,1081.4168 162.402282,1081.46747 162.412282,1081.5168 C162.429615,1081.61613 162.472282,1081.7108 162.536282,1081.7868 C162.600282,1081.86413 162.683615,1081.9248 162.777615,1081.9608 C162.823615,1081.97947 162.872948,1081.99013 162.922282,1081.99613 C162.946948,1081.99813 162.974948,1082.00013 162.994948,1082.00013 L163.052948,1082.00013 L163.516282,1081.9988 L166.296282,1081.99347 L168.149615,1081.99013 L169.076282,1081.98813 L169.192282,1081.98813 C169.232948,1081.98747 169.262948,1081.98947 169.316282,1081.98613 C169.414282,1081.97947 169.510282,1081.94613 169.590948,1081.89147 C169.672282,1081.83747 169.740282,1081.76147 169.783615,1081.67413 C169.806282,1081.63013 169.822282,1081.58347 169.832282,1081.53613 C169.837615,1081.50947 169.839615,1081.4908 169.842948,1081.47213 L169.851615,1081.41547 L169.978948,1080.58013 L171.523615,1080.58013 C171.557615,1080.58013 171.590282,1080.5788 171.623615,1080.57613 C171.689615,1080.57147 171.755615,1080.56213 171.818948,1080.54747 C171.850948,1080.54013 171.882282,1080.53213 171.913615,1080.5228 C172.037615,1080.4848 172.155615,1080.42947 172.263615,1080.35947 C172.290282,1080.34213 172.316948,1080.3228 172.342282,1080.30347 C172.444948,1080.22547 172.536282,1080.13213 172.613615,1080.0268 C172.652282,1079.97413 172.686948,1079.91813 172.718282,1079.85947 C172.732948,1079.82947 172.748282,1079.79947 172.760948,1079.7688 C172.803615,1079.67147 172.848282,1079.56613 172.894948,1079.45747 C172.949615,1079.33013 173.005615,1079.1968 173.062282,1079.05813 C173.112282,1078.93813 173.162282,1078.81613 173.212282,1078.6908 C173.307615,1078.45347 173.402282,1078.2088 173.492282,1077.96613 C173.967615,1076.68213 174.204948,1074.97147 174.204948,1074.97147 C174.319615,1074.91813 174.431615,1074.86213 174.542948,1074.8048 L174.636948,1074.75613 C174.722282,1074.7108 174.806282,1074.6648 174.890282,1074.61813 C174.917615,1074.60213 174.945615,1074.58747 174.972948,1074.57147 C175.195615,1074.44347 175.412282,1074.30813 175.622282,1074.16547 C175.635615,1074.15613 175.648282,1074.14747 175.662282,1074.13813 C175.759615,1074.0708 175.856282,1074.0028 175.950282,1073.93213 C175.960282,1073.92613 175.970282,1073.9188 175.978948,1073.91147 C179.094948,1071.5948 180.684282,1067.62613 179.729615,1063.29747 C179.727615,1063.2888 179.726948,1063.2808 179.724948,1063.27213 C179.703615,1063.18013 179.681615,1063.08813 179.658948,1062.99547 C179.654282,1062.97613 179.649615,1062.95613 179.644282,1062.9368 C179.623615,1062.85147 179.600282,1062.76613 179.576948,1062.6808 C179.569615,1062.65413 179.562282,1062.6268 179.554948,1062.60013 C179.532948,1062.52013 179.508282,1062.44013 179.484282,1062.36013 C179.474948,1062.32747 179.464948,1062.29413 179.454282,1062.26147 C179.430948,1062.18547 179.405615,1062.10947 179.380948,1062.03347 C179.368282,1061.9968 179.356282,1061.95947 179.343615,1061.9228 C179.318282,1061.84947 179.292282,1061.7768 179.266282,1061.70347 C179.252282,1061.66347 179.236948,1061.62213 179.222282,1061.58213 C179.196282,1061.51147 179.168948,1061.44147 179.141615,1061.3708 C179.124282,1061.32747 179.106948,1061.2848 179.089615,1061.2408 C179.061615,1061.1728 179.033615,1061.10413 179.005615,1061.03613 L178.946948,1060.8988 C178.918282,1060.8328 178.888282,1060.76613 178.858282,1060.69947 C178.836948,1060.65147 178.814282,1060.60413 178.792282,1060.55747 C178.762282,1060.49213 178.731615,1060.4268 178.700282,1060.36213 C178.676282,1060.3128 178.651615,1060.26347 178.627615,1060.21413 C178.595615,1060.15013 178.563615,1060.0868 178.531615,1060.02347 C178.505615,1059.9728 178.477615,1059.92213 178.450948,1059.8708 C178.418282,1059.8088 178.384948,1059.74613 178.350948,1059.68347 C178.322282,1059.63213 178.292948,1059.58013 178.263615,1059.52813 C178.228948,1059.4668 178.194948,1059.40547 178.159615,1059.34413 C178.128282,1059.2908 178.096282,1059.23747 178.064282,1059.18413 C178.028948,1059.12413 177.992948,1059.0648 177.956282,1059.0048 C177.922282,1058.95013 177.888282,1058.89547 177.853615,1058.8408 C177.816282,1058.78213 177.780282,1058.72347 177.741615,1058.66413 C177.705615,1058.60813 177.668282,1058.5528 177.630948,1058.4968 L177.626948,1058.4908 L177.919615,1058.5228 C178.022282,1058.53213 178.124948,1058.53613 178.227615,1058.54213 C179.872282,1058.6528 181.534948,1058.52547 183.148948,1058.17947 C183.958282,1058.0128 184.754948,1057.7868 185.536282,1057.52147 C186.318282,1057.2568 187.086282,1056.94813 187.835615,1056.6008 C188.584948,1056.25347 189.315615,1055.8668 190.026948,1055.44413 C190.382282,1055.2328 190.731615,1055.01213 191.074948,1054.7808 C191.418948,1054.54813 191.754282,1054.3108 192.082282,1054.04547 C192.116282,1054.01813 192.138282,1053.97747 192.140282,1053.9308 C192.144282,1053.84413 192.077615,1053.77147 191.990948,1053.76747 L191.988282,1053.76747 L191.101615,1053.72813 C191.774948,1052.4288 192.294282,1051.0488 192.623615,1049.6208 L192.736282,1049.0508 C192.754282,1048.95547 192.774948,1048.86147 192.790282,1048.76547 L192.831615,1048.47813 C192.856948,1048.28613 192.890282,1048.09413 192.908948,1047.90147 L192.960282,1047.3228 C193.010282,1046.55013 193.017615,1045.7728 192.956948,1045.00013'
-            id='Fill-1'
-          ></path>
-        </g>
-      </g>
-    </svg>
-  </div>
-);
-Common.propTypes = CommonPropTypes;
-
-const UncommonPropTypes = { color: CorgiType.color };
-
-export const Uncommon = ({ color }) => (
-  <div>
-    <svg
-      width='96px'
-      height='48px'
-      viewBox='0 0 96 48'
-      version='1.1'
-      xmlns='http://www.w3.org/2000/svg'
-      xlink='http://www.w3.org/1999/xlink'
-    >
-      <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-        <g id='result' transform='translate(-121.000000, -1103.000000)' fill={color}>
-          <path
-            d='M139.074948,1148.0508 C139.088948,1148.01813 139.103615,1147.98547 139.118282,1147.9528 L139.120282,1147.9488 C139.135615,1147.91547 139.150948,1147.8808 139.167615,1147.8468 C139.167615,1147.84613 139.167615,1147.84547 139.168282,1147.8448 C139.183615,1147.81013 139.200282,1147.7748 139.217615,1147.73813 C139.344948,1147.4648 139.494948,1147.1608 139.663615,1146.84547 C139.637615,1147.10213 139.588948,1147.35813 139.519615,1147.60813 C139.446948,1147.8708 139.352948,1148.1348 139.241615,1148.38013 C139.140948,1148.3708 139.038948,1148.3668 138.937615,1148.36547 C138.978282,1148.26947 139.024282,1148.1648 139.074948,1148.0508 M216.956948,1114.00013 C216.833615,1112.45547 216.507615,1110.92547 215.969615,1109.47347 C215.703615,1108.74613 215.390282,1108.0368 215.033615,1107.34947 C214.857615,1107.00547 214.666282,1106.66813 214.464948,1106.3368 C214.262282,1106.0048 214.054282,1105.6828 213.813615,1105.36547 C213.783615,1105.3248 213.735615,1105.29813 213.681615,1105.2968 C213.586948,1105.29413 213.506948,1105.36947 213.504948,1105.46413 L213.504948,1105.47147 C213.502948,1105.5288 213.496948,1105.6028 213.489615,1105.6708 C213.483615,1105.73947 213.472948,1105.81013 213.464282,1105.88013 C213.445615,1106.02013 213.420948,1106.1608 213.394948,1106.3008 C213.342282,1106.58147 213.280282,1106.86147 213.212948,1107.1408 C213.143615,1107.41947 213.072282,1107.69813 212.992282,1107.9748 C212.916948,1108.2528 212.829615,1108.52747 212.745615,1108.80347 C212.401615,1109.90413 212.004948,1110.99013 211.586282,1112.0668 L211.266948,1112.87347 L210.935615,1113.67547 C210.720282,1114.21147 210.485615,1114.74013 210.258282,1115.2728 C209.808948,1116.33147 209.242948,1117.3348 208.526948,1118.2348 C207.818282,1119.1388 206.996948,1119.95347 206.102282,1120.67747 C205.648282,1121.0328 205.187615,1121.37947 204.702282,1121.69147 C204.224282,1122.0128 203.726282,1122.30613 203.221615,1122.58613 C202.714948,1122.86347 202.196282,1123.1188 201.671615,1123.36013 C201.144282,1123.59613 200.609615,1123.8168 200.066948,1124.01947 C199.714948,1124.15147 199.358282,1124.27347 199.000282,1124.38947 C198.933615,1124.3288 198.866948,1124.27013 198.800282,1124.21147 C198.767615,1124.18213 198.734948,1124.15213 198.701615,1124.1228 C198.637615,1124.06747 198.572948,1124.01413 198.508282,1123.95947 C198.468948,1123.92613 198.428948,1123.89213 198.388948,1123.85947 C198.326948,1123.80813 198.263615,1123.7588 198.200948,1123.7088 C198.158282,1123.67547 198.115615,1123.64013 198.072282,1123.60613 C198.010282,1123.5588 197.948948,1123.51213 197.886948,1123.46547 C197.841615,1123.43147 197.796948,1123.39813 197.751615,1123.36413 C197.690282,1123.3188 197.628948,1123.27613 197.566948,1123.23147 C197.520282,1123.1988 197.474282,1123.16547 197.426948,1123.1328 C197.366282,1123.09013 197.304282,1123.0488 197.242282,1123.00813 C197.194948,1122.97547 197.146948,1122.9428 197.098948,1122.91147 C197.037615,1122.8708 196.976282,1122.83213 196.914282,1122.7928 C196.865615,1122.76213 196.816948,1122.7308 196.768282,1122.7008 C196.706282,1122.6628 196.644282,1122.62547 196.582948,1122.5888 C196.533615,1122.5588 196.483615,1122.5288 196.433615,1122.49947 C196.371615,1122.46347 196.310282,1122.42813 196.248282,1122.39347 C196.197615,1122.3648 196.146948,1122.33613 196.096282,1122.30813 C196.034282,1122.27413 195.972282,1122.2408 195.909615,1122.2068 C195.858948,1122.18013 195.808282,1122.1528 195.756948,1122.1268 C195.694948,1122.09413 195.632282,1122.0628 195.568948,1122.03147 C195.517615,1122.00547 195.466948,1121.97947 195.415615,1121.95413 C195.352282,1121.92413 195.288948,1121.89413 195.226948,1121.86413 C195.174282,1121.84013 195.122948,1121.8148 195.070948,1121.7908 C195.007615,1121.76213 194.944948,1121.73413 194.881615,1121.70547 C194.829615,1121.6828 194.777615,1121.65947 194.724948,1121.6368 L194.534948,1121.5568 C194.482282,1121.53547 194.429615,1121.5128 194.376948,1121.49147 C194.313615,1121.46613 194.250282,1121.4408 194.186948,1121.41613 C194.134282,1121.39547 194.081615,1121.3748 194.028948,1121.3548 C193.964948,1121.3308 193.900948,1121.30747 193.837615,1121.28413 C193.783615,1121.26413 193.730948,1121.24547 193.677615,1121.2268 C193.614282,1121.20347 193.550282,1121.18147 193.486282,1121.15947 C193.432948,1121.14213 193.380282,1121.12347 193.326948,1121.10613 C193.262948,1121.0848 193.198948,1121.0648 193.134948,1121.04347 C193.081615,1121.0268 193.028282,1121.01013 192.974948,1120.99347 C192.910948,1120.97413 192.846282,1120.95413 192.782282,1120.93547 C192.728948,1120.92013 192.675615,1120.90413 192.622282,1120.88947 C192.558282,1120.8708 192.494282,1120.8528 192.429615,1120.83547 C192.376282,1120.82013 192.322948,1120.80613 192.269615,1120.79213 C192.204948,1120.77547 192.140948,1120.7588 192.076948,1120.74213 C192.023615,1120.7288 191.970282,1120.71547 191.917615,1120.7028 C191.852282,1120.6868 191.787615,1120.67147 191.723615,1120.65613 C191.670948,1120.64413 191.617615,1120.63147 191.564282,1120.62013 L191.370948,1120.5768 C191.318282,1120.56547 191.264948,1120.55547 191.212282,1120.54413 C191.147615,1120.5308 191.083615,1120.51747 191.018948,1120.50547 C190.966282,1120.4948 190.913615,1120.4848 190.860948,1120.47547 C190.796282,1120.4628 190.731615,1120.45147 190.667615,1120.43947 C190.615615,1120.43013 190.562282,1120.42213 190.510282,1120.4128 C190.446282,1120.40147 190.381615,1120.3908 190.316948,1120.3808 C190.264948,1120.37213 190.212948,1120.3648 190.161615,1120.3568 C190.096948,1120.3468 190.032282,1120.33747 189.967615,1120.32813 C189.916282,1120.32013 189.864948,1120.31413 189.812948,1120.3068 C189.748948,1120.29813 189.684282,1120.2888 189.620282,1120.28147 C189.568948,1120.2748 189.518282,1120.2688 189.466948,1120.2628 C189.402282,1120.25547 189.338282,1120.24747 189.273615,1120.24013 C189.224282,1120.23413 189.174282,1120.23013 189.124282,1120.2248 C189.059615,1120.21813 188.994282,1120.2108 188.928948,1120.20413 C188.880948,1120.20013 188.833615,1120.19613 188.784948,1120.19213 C188.718948,1120.18613 188.652948,1120.18013 188.586948,1120.1748 L188.448282,1120.1648 C188.380948,1120.15947 188.313615,1120.15413 188.247615,1120.14947 C188.202282,1120.14747 188.158282,1120.1448 188.114282,1120.14213 C188.046282,1120.13813 187.977615,1120.13347 187.910282,1120.13013 C187.869615,1120.12813 187.830282,1120.1268 187.789615,1120.1248 C187.718282,1120.12147 187.646282,1120.11813 187.575615,1120.11547 C187.548948,1120.1148 187.522282,1120.11413 187.496282,1120.1128 C187.220948,1120.10347 186.948948,1120.09813 186.680282,1120.09747 C186.652948,1120.09747 186.625615,1120.0968 186.598948,1120.0968 C186.531615,1120.0968 186.464948,1120.09747 186.397615,1120.09813 C186.357615,1120.0988 186.316948,1120.0988 186.277615,1120.09947 C186.215615,1120.10013 186.154282,1120.10147 186.092948,1120.1028 C186.048948,1120.10347 186.004282,1120.10413 185.959615,1120.10547 C185.901615,1120.1068 185.844282,1120.1088 185.786948,1120.1108 C185.739615,1120.11213 185.692948,1120.11347 185.645615,1120.1148 C185.588948,1120.11747 185.533615,1120.11947 185.478282,1120.12147 C185.430948,1120.12347 185.384282,1120.12547 185.337615,1120.12747 C185.282282,1120.13013 185.227615,1120.13347 185.172282,1120.13613 C185.125615,1120.1388 185.078948,1120.14147 185.032948,1120.14413 C184.978282,1120.14747 184.925615,1120.15013 184.872282,1120.15347 C184.825615,1120.1568 184.779615,1120.15947 184.732948,1120.1628 L184.576948,1120.1748 C184.530948,1120.17747 184.484282,1120.1808 184.438282,1120.1848 C184.387615,1120.1888 184.337615,1120.19347 184.286948,1120.19747 C184.240948,1120.20147 184.194948,1120.2048 184.148948,1120.2088 C184.099615,1120.21347 184.050948,1120.21813 184.002282,1120.2228 C183.956282,1120.22747 183.910948,1120.23147 183.865615,1120.23547 C183.816948,1120.2408 183.769615,1120.24613 183.721615,1120.2508 C183.676948,1120.25547 183.632282,1120.26013 183.588282,1120.2648 C183.540282,1120.27013 183.494282,1120.27547 183.447615,1120.2808 C183.402948,1120.28613 183.359615,1120.2908 183.314948,1120.29613 C183.270282,1120.30147 183.224948,1120.3068 183.179615,1120.3128 C183.136282,1120.31747 183.092948,1120.3228 183.049615,1120.3288 C183.005615,1120.33413 182.960948,1120.34013 182.917615,1120.34547 C182.874948,1120.35147 182.832948,1120.3568 182.790948,1120.3628 C182.747615,1120.3688 182.704282,1120.3748 182.661615,1120.3808 C182.620282,1120.3868 182.578948,1120.3928 182.538282,1120.39813 C182.496282,1120.40413 182.454948,1120.4108 182.413615,1120.4168 C182.372948,1120.42347 182.332948,1120.4288 182.292948,1120.4348 C182.252282,1120.44147 182.211615,1120.44813 182.171615,1120.45347 C182.132282,1120.46013 182.092948,1120.4668 182.054948,1120.4728 C182.014948,1120.47947 181.976282,1120.48547 181.937615,1120.49213 C181.898948,1120.4988 181.860948,1120.50547 181.822948,1120.5108 C181.784948,1120.51747 181.748282,1120.52413 181.711615,1120.5308 C181.672948,1120.5368 181.635615,1120.54347 181.598282,1120.5508 C181.562282,1120.5568 181.527615,1120.56347 181.492948,1120.56947 C181.455615,1120.57613 181.419615,1120.5828 181.382948,1120.59013 C181.348282,1120.59613 181.313615,1120.6028 181.280282,1120.6088 C181.245615,1120.61613 181.210948,1120.62213 181.176282,1120.6288 L181.076948,1120.64813 C181.042948,1120.6548 181.009615,1120.66213 180.976948,1120.6688 C180.944948,1120.67547 180.913615,1120.6808 180.882282,1120.68747 C180.849615,1120.69413 180.817615,1120.70147 180.785615,1120.70747 C180.755615,1120.71347 180.726282,1120.72013 180.696948,1120.7268 C180.664948,1120.7328 180.634282,1120.73947 180.602948,1120.74613 C180.574948,1120.7528 180.546948,1120.7588 180.518948,1120.76413 C180.488948,1120.77147 180.460282,1120.77747 180.430948,1120.7848 C180.412282,1120.78813 180.392948,1120.79213 180.373615,1120.7968 C179.396948,1121.0168 178.399615,1121.13147 177.398282,1121.13147 L150.477615,1121.13147 C147.060948,1119.87413 144.922948,1121.28147 144.922948,1121.28147 C143.187615,1118.9668 140.982282,1116.1848 138.698282,1113.8248 C138.632282,1113.75747 138.566948,1113.6888 138.501615,1113.62147 C139.114948,1112.22013 138.644282,1110.1108 138.644282,1110.1108 C139.160948,1108.62213 139.616282,1104.18613 137.489615,1103.1528 C135.362948,1102.12013 134.207615,1106.6168 133.582282,1107.80147 C132.956282,1108.9868 132.394282,1108.8048 132.394282,1108.8048 C132.570948,1109.13147 132.496282,1109.45747 132.397615,1109.67947 C131.870282,1109.6868 131.273615,1109.72613 130.614282,1109.74813 C130.067615,1109.76613 129.478282,1109.77213 128.847615,1109.7388 C128.575615,1109.7248 128.296282,1109.70347 128.008282,1109.67213 C127.980282,1109.64613 127.950282,1109.6168 127.916948,1109.58613 C127.852282,1109.52613 127.778282,1109.4568 127.696948,1109.3808 C127.685615,1109.37013 127.672948,1109.35947 127.662282,1109.3488 C127.632948,1109.32147 127.603615,1109.29347 127.572948,1109.26413 C127.557615,1109.24947 127.540282,1109.23347 127.524948,1109.21813 C127.496948,1109.19213 127.469615,1109.16547 127.440948,1109.13747 C127.422282,1109.12013 127.403615,1109.10147 127.385615,1109.08347 C127.357615,1109.05613 127.329615,1109.02813 127.300948,1108.99947 C127.281615,1108.9808 127.262282,1108.9608 127.242948,1108.94147 C127.213615,1108.91147 127.184282,1108.8808 127.154282,1108.85013 L127.097615,1108.79147 C127.064948,1108.75747 127.032282,1108.72213 126.999615,1108.68613 C126.982948,1108.66947 126.967615,1108.65213 126.950948,1108.6348 C126.902282,1108.58147 126.853615,1108.5268 126.805615,1108.4708 C126.803615,1108.4688 126.802282,1108.4668 126.800282,1108.4648 C126.754282,1108.41147 126.708282,1108.3568 126.662948,1108.30147 C126.646948,1108.2828 126.632948,1108.26347 126.617615,1108.2448 C126.586282,1108.20547 126.554948,1108.16613 126.524282,1108.12547 C126.506282,1108.1028 126.489615,1108.08013 126.472282,1108.0568 C126.444948,1108.01947 126.417615,1107.98213 126.390948,1107.9448 C126.372948,1107.92013 126.356282,1107.89547 126.338948,1107.87013 C126.312948,1107.83347 126.288282,1107.79547 126.264282,1107.75813 C126.248282,1107.73213 126.231615,1107.7068 126.216282,1107.68147 C126.192282,1107.64347 126.169615,1107.60413 126.146282,1107.5648 C126.132282,1107.54013 126.117615,1107.51547 126.104282,1107.4908 C126.080948,1107.44813 126.059615,1107.4048 126.037615,1107.36213 C126.026948,1107.3408 126.015615,1107.3188 126.005615,1107.29747 C125.975615,1107.2328 125.946948,1107.16813 125.921615,1107.1028 C125.366282,1105.67547 123.603615,1102.9708 122.175615,1103.57813 C120.747615,1104.18613 121.020948,1107.83213 121.020948,1107.83213 C121.020948,1107.83213 121.222282,1109.07813 121,1109.38147 C121,1109.38147 121.385615,1109.74613 121.537615,1110.50613 C121.669615,1111.16547 121.247615,1113.33947 122.320948,1114.25613 C122.071615,1114.7748 121.936948,1115.12013 121.936948,1115.12013 C121.936948,1115.12013 122.266948,1116.3908 121.172948,1116.7148 C121.172948,1116.7148 122.590948,1116.7148 122.388282,1117.5648 C122.185615,1118.41613 121.780948,1120.2568 121.982948,1121.28147 C121.982948,1121.28147 121.989615,1121.27213 122.000948,1121.25813 C122.012282,1121.24213 122.030282,1121.22013 122.052282,1121.19413 C122.064948,1121.17947 122.078948,1121.16347 122.094282,1121.14747 C122.112282,1121.12747 122.131615,1121.10813 122.152948,1121.08813 C122.164948,1121.07613 122.177615,1121.0648 122.190948,1121.05347 C122.243615,1121.0068 122.303615,1120.96213 122.366282,1120.93147 C122.398948,1120.91547 122.433615,1120.9028 122.468282,1120.89613 C122.487615,1120.89147 122.506948,1120.8888 122.526948,1120.88813 C122.546948,1120.88747 122.566948,1120.89013 122.587615,1120.89413 C122.593615,1120.89547 122.600948,1120.89613 122.607615,1120.89813 C122.630282,1120.90413 122.652948,1120.91547 122.674948,1120.9288 C122.678948,1120.9308 122.682948,1120.9328 122.686282,1120.93547 C122.691615,1120.9388 122.696282,1120.9428 122.700282,1120.9468 C122.708948,1120.9528 122.716948,1120.95813 122.724948,1120.96547 C122.726948,1120.9668 122.727615,1120.9688 122.729615,1120.9708 C122.753615,1120.9928 122.776282,1121.01947 122.798282,1121.0528 C122.816948,1121.0808 122.834948,1121.11413 122.852282,1121.1508 C122.856948,1121.16013 122.861615,1121.1688 122.865615,1121.17813 C122.866948,1121.18213 122.868948,1121.18747 122.870948,1121.19213 C122.900282,1121.2628 122.927615,1121.34747 122.951615,1121.44947 C122.963615,1121.49813 122.971615,1121.54747 122.979615,1121.5968 C122.984948,1121.6328 122.990282,1121.6688 122.994948,1121.70413 C123.014948,1121.8788 123.018282,1122.0548 123.003615,1122.23013 C122.960948,1122.7268 122.885615,1123.88613 122.935615,1125.42413 L122.935615,1125.4248 C122.936282,1125.43147 122.936282,1125.43947 122.936948,1125.44747 C122.938948,1125.52013 122.942282,1125.59347 122.944948,1125.6688 L122.948948,1125.7568 C122.950948,1125.8108 122.952948,1125.8648 122.955615,1125.91947 C122.957615,1125.95347 122.959615,1125.98747 122.960948,1126.02213 C122.964282,1126.07347 122.966948,1126.1248 122.969615,1126.1768 C122.971615,1126.2128 122.974282,1126.24947 122.976282,1126.28547 C122.978948,1126.33747 122.982282,1126.38947 122.985615,1126.44213 C122.988282,1126.4788 122.990948,1126.51547 122.992948,1126.5528 C122.996948,1126.60613 123.000948,1126.66013 123.004948,1126.71413 C123.006948,1126.7508 123.010282,1126.7868 123.012948,1126.82347 C123.017615,1126.8828 123.022948,1126.94347 123.027615,1127.00347 C123.030282,1127.03547 123.032948,1127.06547 123.034948,1127.09747 C123.043615,1127.19013 123.052282,1127.2828 123.060948,1127.37613 C123.060948,1127.3788 123.060948,1127.38213 123.061615,1127.3848 C123.070282,1127.47613 123.079615,1127.56747 123.088948,1127.65947 C123.092282,1127.6908 123.096282,1127.7228 123.099615,1127.75413 C123.106948,1127.81813 123.114282,1127.8828 123.120948,1127.94813 C123.125615,1127.9848 123.130282,1128.0228 123.134948,1128.0608 C123.142282,1128.12013 123.148948,1128.18013 123.156948,1128.2408 C123.161615,1128.2808 123.166948,1128.32013 123.172282,1128.36013 C123.179615,1128.41947 123.187615,1128.47947 123.196282,1128.5388 C123.200948,1128.57947 123.206948,1128.62013 123.212282,1128.66013 C123.220948,1128.72147 123.229615,1128.78347 123.238948,1128.84413 C123.244948,1128.88347 123.250948,1128.9228 123.256282,1128.96213 C123.266948,1129.02947 123.277615,1129.0968 123.288282,1129.1648 C123.293615,1129.1988 123.298948,1129.23147 123.304282,1129.26547 C123.320282,1129.36747 123.338282,1129.46947 123.355615,1129.57213 C123.355615,1129.57213 124.252282,1136.40347 127.106948,1139.1188 C127.106948,1139.1188 126.925615,1138.65613 126.978282,1138.34413 C131.507615,1144.41147 134.722948,1144.68413 134.722948,1144.68413 L136.100948,1147.25147 L136.052282,1147.49547 C135.956948,1147.51213 135.856282,1147.5328 135.751615,1147.5588 C135.490948,1147.62147 135.208282,1147.71547 134.943615,1147.85347 C134.890948,1147.8808 134.838948,1147.91013 134.786948,1147.94147 C134.696282,1147.99747 134.608282,1148.06013 134.524948,1148.12813 C134.426282,1148.2088 134.334948,1148.29813 134.253615,1148.39813 C134.092282,1148.59413 133.968948,1148.82947 133.904282,1149.11213 C133.862948,1149.29147 133.956282,1149.4648 134.109615,1149.54013 C134.160282,1149.56547 134.218282,1149.58013 134.279615,1149.58013 L136.833615,1149.58013 C136.791615,1149.67547 136.755615,1149.77147 136.726948,1149.87013 C136.707615,1149.93547 136.691615,1150.00213 136.677615,1150.0688 L136.668282,1150.11947 C136.665615,1150.13547 136.662282,1150.1528 136.659615,1150.17613 C136.653615,1150.21947 136.654948,1150.26413 136.660282,1150.30813 C136.682282,1150.4828 136.800282,1150.64147 136.962282,1150.71347 C137.002948,1150.73147 137.044948,1150.74547 137.088948,1150.75213 C137.130282,1150.76013 137.180948,1150.75947 137.208282,1150.75947 L137.792282,1150.7588 L139.349615,1150.75413 L142.464948,1150.74547 L142.659615,1150.7448 L142.756948,1150.74413 C142.790948,1150.74413 142.816948,1150.74547 142.864282,1150.7428 C143.032948,1150.73013 143.194282,1150.62747 143.272282,1150.47213 C143.290948,1150.43547 143.306948,1150.3868 143.312948,1150.35747 C143.321615,1150.32413 143.330948,1150.2908 143.337615,1150.25813 C143.352282,1150.1908 143.364282,1150.12413 143.372948,1150.05747 C143.392282,1149.92347 143.400948,1149.7888 143.404282,1149.6548 C143.408282,1149.13947 143.332948,1148.63813 143.204948,1148.1428 C143.716282,1147.62547 144.106282,1147.0028 144.391615,1146.34413 C144.546282,1145.9868 144.672282,1145.6188 144.776282,1145.24413 C145.059615,1146.09213 144.690948,1146.93747 144.690948,1146.93747 C146.686948,1146.6048 148.124282,1145.25347 148.124282,1145.25347 C148.930948,1145.53813 149.314282,1145.53947 149.314282,1145.53947 C149.932282,1145.15947 150.477615,1145.25347 150.477615,1145.25347 L177.338282,1145.25347 C178.170282,1145.25347 177.615615,1145.80813 177.615615,1145.80813 C179.468948,1145.80813 180.906282,1144.2348 180.906282,1144.2348 L180.666948,1144.96947 C183.565615,1143.78147 183.803615,1143.87613 183.803615,1143.87613 C183.751615,1144.12547 183.613615,1144.2668 183.613615,1144.2668 C183.714282,1144.22547 183.811615,1144.18213 183.906948,1144.1388 C183.935615,1144.12547 183.962282,1144.11213 183.990948,1144.0988 C184.056282,1144.06813 184.120948,1144.03813 184.183615,1144.00747 C184.215615,1143.99147 184.245615,1143.97613 184.276948,1143.95947 C184.333615,1143.9308 184.390282,1143.9028 184.444282,1143.87347 C184.476948,1143.8568 184.506282,1143.84013 184.537615,1143.82213 C184.588948,1143.7948 184.640282,1143.7668 184.688948,1143.7388 C184.720282,1143.72147 184.748948,1143.70347 184.779615,1143.68613 C184.826282,1143.65813 184.873615,1143.6308 184.918948,1143.60347 C184.948948,1143.5848 184.977615,1143.56613 185.006948,1143.54747 C185.048948,1143.52147 185.091615,1143.4948 185.132282,1143.46813 C185.162948,1143.4488 185.190948,1143.4288 185.220282,1143.40947 C185.256948,1143.38413 185.295615,1143.35947 185.331615,1143.33413 C185.361615,1143.31347 185.389615,1143.2928 185.418282,1143.27213 C185.450948,1143.2488 185.484282,1143.22547 185.515615,1143.20213 C185.545615,1143.17947 185.572948,1143.15813 185.601615,1143.1368 C185.629615,1143.11547 185.658948,1143.0928 185.686282,1143.07147 C185.715615,1143.0488 185.742948,1143.0268 185.770948,1143.00413 C185.794948,1142.98413 185.819615,1142.9648 185.842948,1142.94547 C185.872282,1142.9208 185.900282,1142.8968 185.928282,1142.87213 C185.947615,1142.85613 185.967615,1142.83947 185.985615,1142.8228 C186.015615,1142.7968 186.043615,1142.77147 186.071615,1142.74547 C186.086282,1142.73213 186.101615,1142.7188 186.116282,1142.70547 C186.144948,1142.67747 186.172948,1142.6508 186.200948,1142.62413 C186.211615,1142.61413 186.222948,1142.60347 186.232948,1142.59347 C186.262282,1142.5648 186.290282,1142.5368 186.317615,1142.5088 C186.324282,1142.50213 186.331615,1142.4948 186.337615,1142.48813 C186.366948,1142.45747 186.394282,1142.42813 186.420948,1142.40013 C186.423615,1142.3968 186.426948,1142.39347 186.430282,1142.39013 C186.436282,1142.38347 186.442948,1142.37613 186.448948,1142.37013 C186.756282,1142.7168 187.084948,1143.04613 187.442282,1143.34613 L187.692282,1143.55213 C187.778282,1143.61813 187.868282,1143.6768 187.956282,1143.74013 C188.044948,1143.8028 188.132948,1143.86613 188.228948,1143.91747 C188.323615,1143.97147 188.415615,1144.03013 188.512282,1144.07947 L188.809615,1144.21947 C188.913615,1144.25747 189.014948,1144.2988 189.120282,1144.33147 C189.228282,1144.36013 189.334282,1144.3928 189.446948,1144.40813 L189.530282,1144.42213 L189.616282,1144.4288 C189.634282,1144.43013 189.652282,1144.43147 189.671615,1144.43213 C189.717615,1144.80947 189.752948,1145.19013 189.778948,1145.56947 C189.806948,1145.9868 189.822948,1146.40547 189.818948,1146.82213 C189.813615,1147.23813 189.791615,1147.6548 189.719615,1148.05747 C189.683615,1148.25813 189.634282,1148.45547 189.556948,1148.6328 C189.484282,1148.79813 189.377615,1148.93813 189.248282,1149.00147 C188.943615,1148.94413 188.644948,1148.92213 188.338948,1148.93213 C188.019615,1148.94547 187.694282,1149.00013 187.388948,1149.12947 C187.236282,1149.1948 187.089615,1149.28013 186.958948,1149.38747 C186.828948,1149.49547 186.716282,1149.62547 186.628282,1149.76947 C186.540948,1149.91413 186.473615,1150.07013 186.433615,1150.2308 C186.427615,1150.25147 186.422282,1150.2688 186.416948,1150.29213 C186.412948,1150.3168 186.406948,1150.34147 186.405615,1150.3668 C186.401615,1150.4168 186.402282,1150.46747 186.412282,1150.5168 C186.429615,1150.61613 186.472282,1150.7108 186.536282,1150.7868 C186.600282,1150.86413 186.683615,1150.9248 186.777615,1150.9608 C186.823615,1150.97947 186.872948,1150.99013 186.922282,1150.99613 C186.946948,1150.99813 186.974948,1151.00013 186.994948,1151.00013 L187.052948,1151.00013 L187.516282,1150.9988 L190.296282,1150.99347 L192.149615,1150.99013 L193.076282,1150.98813 L193.192282,1150.98813 C193.232948,1150.98747 193.262948,1150.98947 193.316282,1150.98613 C193.414282,1150.97947 193.510282,1150.94613 193.590948,1150.89147 C193.672282,1150.83747 193.740282,1150.76147 193.783615,1150.67413 C193.806282,1150.63013 193.822282,1150.58347 193.832282,1150.53613 C193.837615,1150.50947 193.839615,1150.4908 193.842948,1150.47213 L193.851615,1150.41547 L193.978948,1149.58013 L195.523615,1149.58013 C195.557615,1149.58013 195.590282,1149.5788 195.623615,1149.57613 C195.689615,1149.57147 195.755615,1149.56213 195.818948,1149.54747 C195.850948,1149.54013 195.882282,1149.53213 195.913615,1149.5228 C196.037615,1149.4848 196.155615,1149.42947 196.263615,1149.35947 C196.290282,1149.34213 196.316948,1149.3228 196.342282,1149.30347 C196.444948,1149.22547 196.536282,1149.13213 196.613615,1149.0268 C196.652282,1148.97413 196.686948,1148.91813 196.718282,1148.85947 C196.732948,1148.82947 196.748282,1148.79947 196.760948,1148.7688 C196.803615,1148.67147 196.848282,1148.56613 196.894948,1148.45747 C196.949615,1148.33013 197.005615,1148.1968 197.062282,1148.05813 C197.112282,1147.93813 197.162282,1147.81613 197.212282,1147.6908 C197.307615,1147.45347 197.402282,1147.2088 197.492282,1146.96613 C197.967615,1145.68213 198.204948,1143.97147 198.204948,1143.97147 C198.319615,1143.91813 198.431615,1143.86213 198.542948,1143.8048 L198.636948,1143.75613 C198.722282,1143.7108 198.806282,1143.6648 198.890282,1143.61813 C198.917615,1143.60213 198.945615,1143.58747 198.972948,1143.57147 C199.195615,1143.44347 199.412282,1143.30813 199.622282,1143.16547 C199.635615,1143.15613 199.648282,1143.14747 199.662282,1143.13813 C199.759615,1143.0708 199.856282,1143.0028 199.950282,1142.93213 C199.960282,1142.92613 199.970282,1142.9188 199.978948,1142.91147 C203.094948,1140.5948 204.684282,1136.62613 203.729615,1132.29747 C203.727615,1132.2888 203.726948,1132.2808 203.724948,1132.27213 C203.703615,1132.18013 203.681615,1132.08813 203.658948,1131.99547 C203.654282,1131.97613 203.649615,1131.95613 203.644282,1131.9368 C203.623615,1131.85147 203.600282,1131.76613 203.576948,1131.6808 C203.569615,1131.65413 203.562282,1131.6268 203.554948,1131.60013 C203.532948,1131.52013 203.508282,1131.44013 203.484282,1131.36013 C203.474948,1131.32747 203.464948,1131.29413 203.454282,1131.26147 C203.430948,1131.18547 203.405615,1131.10947 203.380948,1131.03347 C203.368282,1130.9968 203.356282,1130.95947 203.343615,1130.9228 C203.318282,1130.84947 203.292282,1130.7768 203.266282,1130.70347 C203.252282,1130.66347 203.236948,1130.62213 203.222282,1130.58213 C203.196282,1130.51147 203.168948,1130.44147 203.141615,1130.3708 C203.124282,1130.32747 203.106948,1130.2848 203.089615,1130.2408 C203.061615,1130.1728 203.033615,1130.10413 203.005615,1130.03613 L202.946948,1129.8988 C202.918282,1129.8328 202.888282,1129.76613 202.858282,1129.69947 C202.836948,1129.65147 202.814282,1129.60413 202.792282,1129.55747 C202.762282,1129.49213 202.731615,1129.4268 202.700282,1129.36213 C202.676282,1129.3128 202.651615,1129.26347 202.627615,1129.21413 C202.595615,1129.15013 202.563615,1129.0868 202.531615,1129.02347 C202.505615,1128.9728 202.477615,1128.92213 202.450948,1128.8708 C202.418282,1128.8088 202.384948,1128.74613 202.350948,1128.68347 C202.322282,1128.63213 202.292948,1128.58013 202.263615,1128.52813 C202.228948,1128.4668 202.194948,1128.40547 202.159615,1128.34413 C202.128282,1128.2908 202.096282,1128.23747 202.064282,1128.18413 C202.028948,1128.12413 201.992948,1128.0648 201.956282,1128.0048 C201.922282,1127.95013 201.888282,1127.89547 201.853615,1127.8408 C201.816282,1127.78213 201.780282,1127.72347 201.741615,1127.66413 C201.705615,1127.60813 201.668282,1127.5528 201.630948,1127.4968 L201.626948,1127.4908 L201.919615,1127.5228 C202.022282,1127.53213 202.124948,1127.53613 202.227615,1127.54213 C203.872282,1127.6528 205.534948,1127.52547 207.148948,1127.17947 C207.958282,1127.0128 208.754948,1126.7868 209.536282,1126.52147 C210.318282,1126.2568 211.086282,1125.94813 211.835615,1125.6008 C212.584948,1125.25347 213.315615,1124.8668 214.026948,1124.44413 C214.382282,1124.2328 214.731615,1124.01213 215.074948,1123.7808 C215.418948,1123.54813 215.754282,1123.3108 216.082282,1123.04547 C216.116282,1123.01813 216.138282,1122.97747 216.140282,1122.9308 C216.144282,1122.84413 216.077615,1122.77147 215.990948,1122.76747 L215.988282,1122.76747 L215.101615,1122.72813 C215.774948,1121.4288 216.294282,1120.0488 216.623615,1118.6208 L216.736282,1118.0508 C216.754282,1117.95547 216.774948,1117.86147 216.790282,1117.76547 L216.831615,1117.47813 C216.856948,1117.28613 216.890282,1117.09413 216.908948,1116.90147 L216.960282,1116.3228 C217.010282,1115.55013 217.017615,1114.7728 216.956948,1114.00013'
-            id='Fill-1'
-          ></path>
-        </g>
-      </g>
-    </svg>
-  </div>
-);
-Uncommon.propTypes = UncommonPropTypes;
-
-const RarePropTypes = { color: CorgiType.color };
-
-export const Rare = ({ color }) => (
-  <div>
-    <svg
-      width='120px'
-      height='48px'
-      viewBox='0 0 120 48'
-      version='1.1'
-      xmlns='http://www.w3.org/2000/svg'
-      xlink='http://www.w3.org/1999/xlink'
-    >
-      <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-        <g id='result' transform='translate(-121.000000, -1171.000000)' fill={color}>
-          <path
-            d='M139.074948,1216.0508 C139.088948,1216.01813 139.103615,1215.98547 139.118282,1215.9528 L139.120282,1215.9488 C139.135615,1215.91547 139.150948,1215.8808 139.167615,1215.8468 C139.167615,1215.84613 139.167615,1215.84547 139.168282,1215.8448 C139.183615,1215.81013 139.200282,1215.7748 139.217615,1215.73813 C139.344948,1215.4648 139.494948,1215.1608 139.663615,1214.84547 C139.637615,1215.10213 139.588948,1215.35813 139.519615,1215.60813 C139.446948,1215.8708 139.352948,1216.1348 139.241615,1216.38013 C139.140948,1216.3708 139.038948,1216.3668 138.937615,1216.36547 C138.978282,1216.26947 139.024282,1216.1648 139.074948,1216.0508 M240.956948,1182.00013 C240.833615,1180.45547 240.507615,1178.92547 239.969615,1177.47347 C239.703615,1176.74613 239.390282,1176.0368 239.033615,1175.34947 C238.857615,1175.00547 238.666282,1174.66813 238.464948,1174.3368 C238.262282,1174.0048 238.054282,1173.6828 237.813615,1173.36547 C237.783615,1173.3248 237.735615,1173.29813 237.681615,1173.2968 C237.586948,1173.29413 237.506948,1173.36947 237.504948,1173.46413 L237.504948,1173.47147 C237.502948,1173.5288 237.496948,1173.6028 237.489615,1173.6708 C237.483615,1173.73947 237.472948,1173.81013 237.464282,1173.88013 C237.445615,1174.02013 237.420948,1174.1608 237.394948,1174.3008 C237.342282,1174.58147 237.280282,1174.86147 237.212948,1175.1408 C237.143615,1175.41947 237.072282,1175.69813 236.992282,1175.9748 C236.916948,1176.2528 236.829615,1176.52747 236.745615,1176.80347 C236.401615,1177.90413 236.004948,1178.99013 235.586282,1180.0668 L235.266948,1180.87347 L234.935615,1181.67547 C234.720282,1182.21147 234.485615,1182.74013 234.258282,1183.2728 C233.808948,1184.33147 233.242948,1185.3348 232.526948,1186.2348 C231.818282,1187.1388 230.996948,1187.95347 230.102282,1188.67747 C229.648282,1189.0328 229.187615,1189.37947 228.702282,1189.69147 C228.224282,1190.0128 227.726282,1190.30613 227.221615,1190.58613 C226.714948,1190.86347 226.196282,1191.1188 225.671615,1191.36013 C225.144282,1191.59613 224.609615,1191.8168 224.066948,1192.01947 C223.714948,1192.15147 223.358282,1192.27347 223.000282,1192.38947 C222.933615,1192.3288 222.866948,1192.27013 222.800282,1192.21147 C222.767615,1192.18213 222.734948,1192.15213 222.701615,1192.1228 C222.637615,1192.06747 222.572948,1192.01413 222.508282,1191.95947 C222.468948,1191.92613 222.428948,1191.89213 222.388948,1191.85947 C222.326948,1191.80813 222.263615,1191.7588 222.200948,1191.7088 C222.158282,1191.67547 222.115615,1191.64013 222.072282,1191.60613 C222.010282,1191.5588 221.948948,1191.51213 221.886948,1191.46547 C221.841615,1191.43147 221.796948,1191.39813 221.751615,1191.36413 C221.690282,1191.3188 221.628948,1191.27613 221.566948,1191.23147 C221.520282,1191.1988 221.474282,1191.16547 221.426948,1191.1328 C221.366282,1191.09013 221.304282,1191.0488 221.242282,1191.00813 C221.194948,1190.97547 221.146948,1190.9428 221.098948,1190.91147 C221.037615,1190.8708 220.976282,1190.83213 220.914282,1190.7928 C220.865615,1190.76213 220.816948,1190.7308 220.768282,1190.7008 C220.706282,1190.6628 220.644282,1190.62547 220.582948,1190.5888 C220.533615,1190.5588 220.483615,1190.5288 220.433615,1190.49947 C220.371615,1190.46347 220.310282,1190.42813 220.248282,1190.39347 C220.197615,1190.3648 220.146948,1190.33613 220.096282,1190.30813 C220.034282,1190.27413 219.972282,1190.2408 219.909615,1190.2068 C219.858948,1190.18013 219.808282,1190.1528 219.756948,1190.1268 C219.694948,1190.09413 219.632282,1190.0628 219.568948,1190.03147 C219.517615,1190.00547 219.466948,1189.97947 219.415615,1189.95413 C219.352282,1189.92413 219.288948,1189.89413 219.226948,1189.86413 C219.174282,1189.84013 219.122948,1189.8148 219.070948,1189.7908 C219.007615,1189.76213 218.944948,1189.73413 218.881615,1189.70547 C218.829615,1189.6828 218.777615,1189.65947 218.724948,1189.6368 L218.534948,1189.5568 C218.482282,1189.53547 218.429615,1189.5128 218.376948,1189.49147 C218.313615,1189.46613 218.250282,1189.4408 218.186948,1189.41613 C218.134282,1189.39547 218.081615,1189.3748 218.028948,1189.3548 C217.964948,1189.3308 217.900948,1189.30747 217.837615,1189.28413 C217.783615,1189.26413 217.730948,1189.24547 217.677615,1189.2268 C217.614282,1189.20347 217.550282,1189.18147 217.486282,1189.15947 C217.432948,1189.14213 217.380282,1189.12347 217.326948,1189.10613 C217.262948,1189.0848 217.198948,1189.0648 217.134948,1189.04347 C217.081615,1189.0268 217.028282,1189.01013 216.974948,1188.99347 C216.910948,1188.97413 216.846282,1188.95413 216.782282,1188.93547 C216.728948,1188.92013 216.675615,1188.90413 216.622282,1188.88947 C216.558282,1188.8708 216.494282,1188.8528 216.429615,1188.83547 C216.376282,1188.82013 216.322948,1188.80613 216.269615,1188.79213 C216.204948,1188.77547 216.140948,1188.7588 216.076948,1188.74213 C216.023615,1188.7288 215.970282,1188.71547 215.917615,1188.7028 C215.852282,1188.6868 215.787615,1188.67147 215.723615,1188.65613 C215.670948,1188.64413 215.617615,1188.63147 215.564282,1188.62013 L215.370948,1188.5768 C215.318282,1188.56547 215.264948,1188.55547 215.212282,1188.54413 C215.147615,1188.5308 215.083615,1188.51747 215.018948,1188.50547 C214.966282,1188.4948 214.913615,1188.4848 214.860948,1188.47547 C214.796282,1188.4628 214.731615,1188.45147 214.667615,1188.43947 C214.615615,1188.43013 214.562282,1188.42213 214.510282,1188.4128 C214.446282,1188.40147 214.381615,1188.3908 214.316948,1188.3808 C214.264948,1188.37213 214.212948,1188.3648 214.161615,1188.3568 C214.096948,1188.3468 214.032282,1188.33747 213.967615,1188.32813 C213.916282,1188.32013 213.864948,1188.31413 213.812948,1188.3068 C213.748948,1188.29813 213.684282,1188.2888 213.620282,1188.28147 C213.568948,1188.2748 213.518282,1188.2688 213.466948,1188.2628 C213.402282,1188.25547 213.338282,1188.24747 213.273615,1188.24013 C213.224282,1188.23413 213.174282,1188.23013 213.124282,1188.2248 C213.059615,1188.21813 212.994282,1188.2108 212.928948,1188.20413 C212.880948,1188.20013 212.833615,1188.19613 212.784948,1188.19213 C212.718948,1188.18613 212.652948,1188.18013 212.586948,1188.1748 L212.448282,1188.1648 C212.380948,1188.15947 212.313615,1188.15413 212.247615,1188.14947 C212.202282,1188.14747 212.158282,1188.1448 212.114282,1188.14213 C212.046282,1188.13813 211.977615,1188.13347 211.910282,1188.13013 C211.869615,1188.12813 211.830282,1188.1268 211.789615,1188.1248 C211.718282,1188.12147 211.646282,1188.11813 211.575615,1188.11547 C211.548948,1188.1148 211.522282,1188.11413 211.496282,1188.1128 C211.220948,1188.10347 210.948948,1188.09813 210.680282,1188.09747 C210.652948,1188.09747 210.625615,1188.0968 210.598948,1188.0968 C210.531615,1188.0968 210.464948,1188.09747 210.397615,1188.09813 C210.357615,1188.0988 210.316948,1188.0988 210.277615,1188.09947 C210.215615,1188.10013 210.154282,1188.10147 210.092948,1188.1028 C210.048948,1188.10347 210.004282,1188.10413 209.959615,1188.10547 C209.901615,1188.1068 209.844282,1188.1088 209.786948,1188.1108 C209.739615,1188.11213 209.692948,1188.11347 209.645615,1188.1148 C209.588948,1188.11747 209.533615,1188.11947 209.478282,1188.12147 C209.430948,1188.12347 209.384282,1188.12547 209.337615,1188.12747 C209.282282,1188.13013 209.227615,1188.13347 209.172282,1188.13613 C209.125615,1188.1388 209.078948,1188.14147 209.032948,1188.14413 C208.978282,1188.14747 208.925615,1188.15013 208.872282,1188.15347 C208.825615,1188.1568 208.779615,1188.15947 208.732948,1188.1628 L208.576948,1188.1748 C208.530948,1188.17747 208.484282,1188.1808 208.438282,1188.1848 C208.387615,1188.1888 208.337615,1188.19347 208.286948,1188.19747 C208.240948,1188.20147 208.194948,1188.2048 208.148948,1188.2088 C208.099615,1188.21347 208.050948,1188.21813 208.002282,1188.2228 C207.956282,1188.22747 207.910948,1188.23147 207.865615,1188.23547 C207.816948,1188.2408 207.769615,1188.24613 207.721615,1188.2508 C207.676948,1188.25547 207.632282,1188.26013 207.588282,1188.2648 C207.540282,1188.27013 207.494282,1188.27547 207.447615,1188.2808 C207.402948,1188.28613 207.359615,1188.2908 207.314948,1188.29613 C207.270282,1188.30147 207.224948,1188.3068 207.179615,1188.3128 C207.136282,1188.31747 207.092948,1188.3228 207.049615,1188.3288 C207.005615,1188.33413 206.960948,1188.34013 206.917615,1188.34547 C206.874948,1188.35147 206.832948,1188.3568 206.790948,1188.3628 C206.747615,1188.3688 206.704282,1188.3748 206.661615,1188.3808 C206.620282,1188.3868 206.578948,1188.3928 206.538282,1188.39813 C206.496282,1188.40413 206.454948,1188.4108 206.413615,1188.4168 C206.372948,1188.42347 206.332948,1188.4288 206.292948,1188.4348 C206.252282,1188.44147 206.211615,1188.44813 206.171615,1188.45347 C206.132282,1188.46013 206.092948,1188.4668 206.054948,1188.4728 C206.014948,1188.47947 205.976282,1188.48547 205.937615,1188.49213 C205.898948,1188.4988 205.860948,1188.50547 205.822948,1188.5108 C205.784948,1188.51747 205.748282,1188.52413 205.711615,1188.5308 C205.672948,1188.5368 205.635615,1188.54347 205.598282,1188.5508 C205.562282,1188.5568 205.527615,1188.56347 205.492948,1188.56947 C205.455615,1188.57613 205.419615,1188.5828 205.382948,1188.59013 C205.348282,1188.59613 205.313615,1188.6028 205.280282,1188.6088 C205.245615,1188.61613 205.210948,1188.62213 205.176282,1188.6288 L205.076948,1188.64813 C205.042948,1188.6548 205.009615,1188.66213 204.976948,1188.6688 C204.944948,1188.67547 204.913615,1188.6808 204.882282,1188.68747 C204.849615,1188.69413 204.817615,1188.70147 204.785615,1188.70747 C204.755615,1188.71347 204.726282,1188.72013 204.696948,1188.7268 C204.664948,1188.7328 204.634282,1188.73947 204.602948,1188.74613 C204.574948,1188.7528 204.546948,1188.7588 204.518948,1188.76413 C204.488948,1188.77147 204.460282,1188.77747 204.430948,1188.7848 C204.412282,1188.78813 204.392948,1188.79213 204.373615,1188.7968 C203.396948,1189.0168 202.399615,1189.13147 201.398282,1189.13147 L150.477615,1189.13147 C147.060948,1187.87413 144.922948,1189.28147 144.922948,1189.28147 C143.187615,1186.9668 140.982282,1184.1848 138.698282,1181.8248 C138.632282,1181.75747 138.566948,1181.6888 138.501615,1181.62147 C139.114948,1180.22013 138.644282,1178.1108 138.644282,1178.1108 C139.160948,1176.62213 139.616282,1172.18613 137.489615,1171.1528 C135.362948,1170.12013 134.207615,1174.6168 133.582282,1175.80147 C132.956282,1176.9868 132.394282,1176.8048 132.394282,1176.8048 C132.570948,1177.13147 132.496282,1177.45747 132.397615,1177.67947 C131.870282,1177.6868 131.273615,1177.72613 130.614282,1177.74813 C130.067615,1177.76613 129.478282,1177.77213 128.847615,1177.7388 C128.575615,1177.7248 128.296282,1177.70347 128.008282,1177.67213 C127.980282,1177.64613 127.950282,1177.6168 127.916948,1177.58613 C127.852282,1177.52613 127.778282,1177.4568 127.696948,1177.3808 C127.685615,1177.37013 127.672948,1177.35947 127.662282,1177.3488 C127.632948,1177.32147 127.603615,1177.29347 127.572948,1177.26413 C127.557615,1177.24947 127.540282,1177.23347 127.524948,1177.21813 C127.496948,1177.19213 127.469615,1177.16547 127.440948,1177.13747 C127.422282,1177.12013 127.403615,1177.10147 127.385615,1177.08347 C127.357615,1177.05613 127.329615,1177.02813 127.300948,1176.99947 C127.281615,1176.9808 127.262282,1176.9608 127.242948,1176.94147 C127.213615,1176.91147 127.184282,1176.8808 127.154282,1176.85013 L127.097615,1176.79147 C127.064948,1176.75747 127.032282,1176.72213 126.999615,1176.68613 C126.982948,1176.66947 126.967615,1176.65213 126.950948,1176.6348 C126.902282,1176.58147 126.853615,1176.5268 126.805615,1176.4708 C126.803615,1176.4688 126.802282,1176.4668 126.800282,1176.4648 C126.754282,1176.41147 126.708282,1176.3568 126.662948,1176.30147 C126.646948,1176.2828 126.632948,1176.26347 126.617615,1176.2448 C126.586282,1176.20547 126.554948,1176.16613 126.524282,1176.12547 C126.506282,1176.1028 126.489615,1176.08013 126.472282,1176.0568 C126.444948,1176.01947 126.417615,1175.98213 126.390948,1175.9448 C126.372948,1175.92013 126.356282,1175.89547 126.338948,1175.87013 C126.312948,1175.83347 126.288282,1175.79547 126.264282,1175.75813 C126.248282,1175.73213 126.231615,1175.7068 126.216282,1175.68147 C126.192282,1175.64347 126.169615,1175.60413 126.146282,1175.5648 C126.132282,1175.54013 126.117615,1175.51547 126.104282,1175.4908 C126.080948,1175.44813 126.059615,1175.4048 126.037615,1175.36213 C126.026948,1175.3408 126.015615,1175.3188 126.005615,1175.29747 C125.975615,1175.2328 125.946948,1175.16813 125.921615,1175.1028 C125.366282,1173.67547 123.603615,1170.9708 122.175615,1171.57813 C120.747615,1172.18613 121.020948,1175.83213 121.020948,1175.83213 C121.020948,1175.83213 121.222282,1177.07813 121,1177.38147 C121,1177.38147 121.385615,1177.74613 121.537615,1178.50613 C121.669615,1179.16547 121.247615,1181.33947 122.320948,1182.25613 C122.071615,1182.7748 121.936948,1183.12013 121.936948,1183.12013 C121.936948,1183.12013 122.266948,1184.3908 121.172948,1184.7148 C121.172948,1184.7148 122.590948,1184.7148 122.388282,1185.5648 C122.185615,1186.41613 121.780948,1188.2568 121.982948,1189.28147 C121.982948,1189.28147 121.989615,1189.27213 122.000948,1189.25813 C122.012282,1189.24213 122.030282,1189.22013 122.052282,1189.19413 C122.064948,1189.17947 122.078948,1189.16347 122.094282,1189.14747 C122.112282,1189.12747 122.131615,1189.10813 122.152948,1189.08813 C122.164948,1189.07613 122.177615,1189.0648 122.190948,1189.05347 C122.243615,1189.0068 122.303615,1188.96213 122.366282,1188.93147 C122.398948,1188.91547 122.433615,1188.9028 122.468282,1188.89613 C122.487615,1188.89147 122.506948,1188.8888 122.526948,1188.88813 C122.546948,1188.88747 122.566948,1188.89013 122.587615,1188.89413 C122.593615,1188.89547 122.600948,1188.89613 122.607615,1188.89813 C122.630282,1188.90413 122.652948,1188.91547 122.674948,1188.9288 C122.678948,1188.9308 122.682948,1188.9328 122.686282,1188.93547 C122.691615,1188.9388 122.696282,1188.9428 122.700282,1188.9468 C122.708948,1188.9528 122.716948,1188.95813 122.724948,1188.96547 C122.726948,1188.9668 122.727615,1188.9688 122.729615,1188.9708 C122.753615,1188.9928 122.776282,1189.01947 122.798282,1189.0528 C122.816948,1189.0808 122.834948,1189.11413 122.852282,1189.1508 C122.856948,1189.16013 122.861615,1189.1688 122.865615,1189.17813 C122.866948,1189.18213 122.868948,1189.18747 122.870948,1189.19213 C122.900282,1189.2628 122.927615,1189.34747 122.951615,1189.44947 C122.963615,1189.49813 122.971615,1189.54747 122.979615,1189.5968 C122.984948,1189.6328 122.990282,1189.6688 122.994948,1189.70413 C123.014948,1189.8788 123.018282,1190.0548 123.003615,1190.23013 C122.960948,1190.7268 122.885615,1191.88613 122.935615,1193.42413 L122.935615,1193.4248 C122.936282,1193.43147 122.936282,1193.43947 122.936948,1193.44747 C122.938948,1193.52013 122.942282,1193.59347 122.944948,1193.6688 L122.948948,1193.7568 C122.950948,1193.8108 122.952948,1193.8648 122.955615,1193.91947 C122.957615,1193.95347 122.959615,1193.98747 122.960948,1194.02213 C122.964282,1194.07347 122.966948,1194.1248 122.969615,1194.1768 C122.971615,1194.2128 122.974282,1194.24947 122.976282,1194.28547 C122.978948,1194.33747 122.982282,1194.38947 122.985615,1194.44213 C122.988282,1194.4788 122.990948,1194.51547 122.992948,1194.5528 C122.996948,1194.60613 123.000948,1194.66013 123.004948,1194.71413 C123.006948,1194.7508 123.010282,1194.7868 123.012948,1194.82347 C123.017615,1194.8828 123.022948,1194.94347 123.027615,1195.00347 C123.030282,1195.03547 123.032948,1195.06547 123.034948,1195.09747 C123.043615,1195.19013 123.052282,1195.2828 123.060948,1195.37613 C123.060948,1195.3788 123.060948,1195.38213 123.061615,1195.3848 C123.070282,1195.47613 123.079615,1195.56747 123.088948,1195.65947 C123.092282,1195.6908 123.096282,1195.7228 123.099615,1195.75413 C123.106948,1195.81813 123.114282,1195.8828 123.120948,1195.94813 C123.125615,1195.9848 123.130282,1196.0228 123.134948,1196.0608 C123.142282,1196.12013 123.148948,1196.18013 123.156948,1196.2408 C123.161615,1196.2808 123.166948,1196.32013 123.172282,1196.36013 C123.179615,1196.41947 123.187615,1196.47947 123.196282,1196.5388 C123.200948,1196.57947 123.206948,1196.62013 123.212282,1196.66013 C123.220948,1196.72147 123.229615,1196.78347 123.238948,1196.84413 C123.244948,1196.88347 123.250948,1196.9228 123.256282,1196.96213 C123.266948,1197.02947 123.277615,1197.0968 123.288282,1197.1648 C123.293615,1197.1988 123.298948,1197.23147 123.304282,1197.26547 C123.320282,1197.36747 123.338282,1197.46947 123.355615,1197.57213 C123.355615,1197.57213 124.252282,1204.40347 127.106948,1207.1188 C127.106948,1207.1188 126.925615,1206.65613 126.978282,1206.34413 C131.507615,1212.41147 134.722948,1212.68413 134.722948,1212.68413 L136.100948,1215.25147 L136.052282,1215.49547 C135.956948,1215.51213 135.856282,1215.5328 135.751615,1215.5588 C135.490948,1215.62147 135.208282,1215.71547 134.943615,1215.85347 C134.890948,1215.8808 134.838948,1215.91013 134.786948,1215.94147 C134.696282,1215.99747 134.608282,1216.06013 134.524948,1216.12813 C134.426282,1216.2088 134.334948,1216.29813 134.253615,1216.39813 C134.092282,1216.59413 133.968948,1216.82947 133.904282,1217.11213 C133.862948,1217.29147 133.956282,1217.4648 134.109615,1217.54013 C134.160282,1217.56547 134.218282,1217.58013 134.279615,1217.58013 L136.833615,1217.58013 C136.791615,1217.67547 136.755615,1217.77147 136.726948,1217.87013 C136.707615,1217.93547 136.691615,1218.00213 136.677615,1218.0688 L136.668282,1218.11947 C136.665615,1218.13547 136.662282,1218.1528 136.659615,1218.17613 C136.653615,1218.21947 136.654948,1218.26413 136.660282,1218.30813 C136.682282,1218.4828 136.800282,1218.64147 136.962282,1218.71347 C137.002948,1218.73147 137.044948,1218.74547 137.088948,1218.75213 C137.130282,1218.76013 137.180948,1218.75947 137.208282,1218.75947 L137.792282,1218.7588 L139.349615,1218.75413 L142.464948,1218.74547 L142.659615,1218.7448 L142.756948,1218.74413 C142.790948,1218.74413 142.816948,1218.74547 142.864282,1218.7428 C143.032948,1218.73013 143.194282,1218.62747 143.272282,1218.47213 C143.290948,1218.43547 143.306948,1218.3868 143.312948,1218.35747 C143.321615,1218.32413 143.330948,1218.2908 143.337615,1218.25813 C143.352282,1218.1908 143.364282,1218.12413 143.372948,1218.05747 C143.392282,1217.92347 143.400948,1217.7888 143.404282,1217.6548 C143.408282,1217.13947 143.332948,1216.63813 143.204948,1216.1428 C143.716282,1215.62547 144.106282,1215.0028 144.391615,1214.34413 C144.546282,1213.9868 144.672282,1213.6188 144.776282,1213.24413 C145.059615,1214.09213 144.690948,1214.93747 144.690948,1214.93747 C146.686948,1214.6048 148.124282,1213.25347 148.124282,1213.25347 C148.930948,1213.53813 149.314282,1213.53947 149.314282,1213.53947 C149.932282,1213.15947 150.477615,1213.25347 150.477615,1213.25347 L201.338282,1213.25347 C202.170282,1213.25347 201.615615,1213.80813 201.615615,1213.80813 C203.468948,1213.80813 204.906282,1212.2348 204.906282,1212.2348 L204.666948,1212.96947 C207.565615,1211.78147 207.803615,1211.87613 207.803615,1211.87613 C207.751615,1212.12547 207.613615,1212.2668 207.613615,1212.2668 C207.714282,1212.22547 207.811615,1212.18213 207.906948,1212.1388 C207.935615,1212.12547 207.962282,1212.11213 207.990948,1212.0988 C208.056282,1212.06813 208.120948,1212.03813 208.183615,1212.00747 C208.215615,1211.99147 208.245615,1211.97613 208.276948,1211.95947 C208.333615,1211.9308 208.390282,1211.9028 208.444282,1211.87347 C208.476948,1211.8568 208.506282,1211.84013 208.537615,1211.82213 C208.588948,1211.7948 208.640282,1211.7668 208.688948,1211.7388 C208.720282,1211.72147 208.748948,1211.70347 208.779615,1211.68613 C208.826282,1211.65813 208.873615,1211.6308 208.918948,1211.60347 C208.948948,1211.5848 208.977615,1211.56613 209.006948,1211.54747 C209.048948,1211.52147 209.091615,1211.4948 209.132282,1211.46813 C209.162948,1211.4488 209.190948,1211.4288 209.220282,1211.40947 C209.256948,1211.38413 209.295615,1211.35947 209.331615,1211.33413 C209.361615,1211.31347 209.389615,1211.2928 209.418282,1211.27213 C209.450948,1211.2488 209.484282,1211.22547 209.515615,1211.20213 C209.545615,1211.17947 209.572948,1211.15813 209.601615,1211.1368 C209.629615,1211.11547 209.658948,1211.0928 209.686282,1211.07147 C209.715615,1211.0488 209.742948,1211.0268 209.770948,1211.00413 C209.794948,1210.98413 209.819615,1210.9648 209.842948,1210.94547 C209.872282,1210.9208 209.900282,1210.8968 209.928282,1210.87213 C209.947615,1210.85613 209.967615,1210.83947 209.985615,1210.8228 C210.015615,1210.7968 210.043615,1210.77147 210.071615,1210.74547 C210.086282,1210.73213 210.101615,1210.7188 210.116282,1210.70547 C210.144948,1210.67747 210.172948,1210.6508 210.200948,1210.62413 C210.211615,1210.61413 210.222948,1210.60347 210.232948,1210.59347 C210.262282,1210.5648 210.290282,1210.5368 210.317615,1210.5088 C210.324282,1210.50213 210.331615,1210.4948 210.337615,1210.48813 C210.366948,1210.45747 210.394282,1210.42813 210.420948,1210.40013 C210.423615,1210.3968 210.426948,1210.39347 210.430282,1210.39013 C210.436282,1210.38347 210.442948,1210.37613 210.448948,1210.37013 C210.756282,1210.7168 211.084948,1211.04613 211.442282,1211.34613 L211.692282,1211.55213 C211.778282,1211.61813 211.868282,1211.6768 211.956282,1211.74013 C212.044948,1211.8028 212.132948,1211.86613 212.228948,1211.91747 C212.323615,1211.97147 212.415615,1212.03013 212.512282,1212.07947 L212.809615,1212.21947 C212.913615,1212.25747 213.014948,1212.2988 213.120282,1212.33147 C213.228282,1212.36013 213.334282,1212.3928 213.446948,1212.40813 L213.530282,1212.42213 L213.616282,1212.4288 C213.634282,1212.43013 213.652282,1212.43147 213.671615,1212.43213 C213.717615,1212.80947 213.752948,1213.19013 213.778948,1213.56947 C213.806948,1213.9868 213.822948,1214.40547 213.818948,1214.82213 C213.813615,1215.23813 213.791615,1215.6548 213.719615,1216.05747 C213.683615,1216.25813 213.634282,1216.45547 213.556948,1216.6328 C213.484282,1216.79813 213.377615,1216.93813 213.248282,1217.00147 C212.943615,1216.94413 212.644948,1216.92213 212.338948,1216.93213 C212.019615,1216.94547 211.694282,1217.00013 211.388948,1217.12947 C211.236282,1217.1948 211.089615,1217.28013 210.958948,1217.38747 C210.828948,1217.49547 210.716282,1217.62547 210.628282,1217.76947 C210.540948,1217.91413 210.473615,1218.07013 210.433615,1218.2308 C210.427615,1218.25147 210.422282,1218.2688 210.416948,1218.29213 C210.412948,1218.3168 210.406948,1218.34147 210.405615,1218.3668 C210.401615,1218.4168 210.402282,1218.46747 210.412282,1218.5168 C210.429615,1218.61613 210.472282,1218.7108 210.536282,1218.7868 C210.600282,1218.86413 210.683615,1218.9248 210.777615,1218.9608 C210.823615,1218.97947 210.872948,1218.99013 210.922282,1218.99613 C210.946948,1218.99813 210.974948,1219.00013 210.994948,1219.00013 L211.052948,1219.00013 L211.516282,1218.9988 L214.296282,1218.99347 L216.149615,1218.99013 L217.076282,1218.98813 L217.192282,1218.98813 C217.232948,1218.98747 217.262948,1218.98947 217.316282,1218.98613 C217.414282,1218.97947 217.510282,1218.94613 217.590948,1218.89147 C217.672282,1218.83747 217.740282,1218.76147 217.783615,1218.67413 C217.806282,1218.63013 217.822282,1218.58347 217.832282,1218.53613 C217.837615,1218.50947 217.839615,1218.4908 217.842948,1218.47213 L217.851615,1218.41547 L217.978948,1217.58013 L219.523615,1217.58013 C219.557615,1217.58013 219.590282,1217.5788 219.623615,1217.57613 C219.689615,1217.57147 219.755615,1217.56213 219.818948,1217.54747 C219.850948,1217.54013 219.882282,1217.53213 219.913615,1217.5228 C220.037615,1217.4848 220.155615,1217.42947 220.263615,1217.35947 C220.290282,1217.34213 220.316948,1217.3228 220.342282,1217.30347 C220.444948,1217.22547 220.536282,1217.13213 220.613615,1217.0268 C220.652282,1216.97413 220.686948,1216.91813 220.718282,1216.85947 C220.732948,1216.82947 220.748282,1216.79947 220.760948,1216.7688 C220.803615,1216.67147 220.848282,1216.56613 220.894948,1216.45747 C220.949615,1216.33013 221.005615,1216.1968 221.062282,1216.05813 C221.112282,1215.93813 221.162282,1215.81613 221.212282,1215.6908 C221.307615,1215.45347 221.402282,1215.2088 221.492282,1214.96613 C221.967615,1213.68213 222.204948,1211.97147 222.204948,1211.97147 C222.319615,1211.91813 222.431615,1211.86213 222.542948,1211.8048 L222.636948,1211.75613 C222.722282,1211.7108 222.806282,1211.6648 222.890282,1211.61813 C222.917615,1211.60213 222.945615,1211.58747 222.972948,1211.57147 C223.195615,1211.44347 223.412282,1211.30813 223.622282,1211.16547 C223.635615,1211.15613 223.648282,1211.14747 223.662282,1211.13813 C223.759615,1211.0708 223.856282,1211.0028 223.950282,1210.93213 C223.960282,1210.92613 223.970282,1210.9188 223.978948,1210.91147 C227.094948,1208.5948 228.684282,1204.62613 227.729615,1200.29747 C227.727615,1200.2888 227.726948,1200.2808 227.724948,1200.27213 C227.703615,1200.18013 227.681615,1200.08813 227.658948,1199.99547 C227.654282,1199.97613 227.649615,1199.95613 227.644282,1199.9368 C227.623615,1199.85147 227.600282,1199.76613 227.576948,1199.6808 C227.569615,1199.65413 227.562282,1199.6268 227.554948,1199.60013 C227.532948,1199.52013 227.508282,1199.44013 227.484282,1199.36013 C227.474948,1199.32747 227.464948,1199.29413 227.454282,1199.26147 C227.430948,1199.18547 227.405615,1199.10947 227.380948,1199.03347 C227.368282,1198.9968 227.356282,1198.95947 227.343615,1198.9228 C227.318282,1198.84947 227.292282,1198.7768 227.266282,1198.70347 C227.252282,1198.66347 227.236948,1198.62213 227.222282,1198.58213 C227.196282,1198.51147 227.168948,1198.44147 227.141615,1198.3708 C227.124282,1198.32747 227.106948,1198.2848 227.089615,1198.2408 C227.061615,1198.1728 227.033615,1198.10413 227.005615,1198.03613 L226.946948,1197.8988 C226.918282,1197.8328 226.888282,1197.76613 226.858282,1197.69947 C226.836948,1197.65147 226.814282,1197.60413 226.792282,1197.55747 C226.762282,1197.49213 226.731615,1197.4268 226.700282,1197.36213 C226.676282,1197.3128 226.651615,1197.26347 226.627615,1197.21413 C226.595615,1197.15013 226.563615,1197.0868 226.531615,1197.02347 C226.505615,1196.9728 226.477615,1196.92213 226.450948,1196.8708 C226.418282,1196.8088 226.384948,1196.74613 226.350948,1196.68347 C226.322282,1196.63213 226.292948,1196.58013 226.263615,1196.52813 C226.228948,1196.4668 226.194948,1196.40547 226.159615,1196.34413 C226.128282,1196.2908 226.096282,1196.23747 226.064282,1196.18413 C226.028948,1196.12413 225.992948,1196.0648 225.956282,1196.0048 C225.922282,1195.95013 225.888282,1195.89547 225.853615,1195.8408 C225.816282,1195.78213 225.780282,1195.72347 225.741615,1195.66413 C225.705615,1195.60813 225.668282,1195.5528 225.630948,1195.4968 L225.626948,1195.4908 L225.919615,1195.5228 C226.022282,1195.53213 226.124948,1195.53613 226.227615,1195.54213 C227.872282,1195.6528 229.534948,1195.52547 231.148948,1195.17947 C231.958282,1195.0128 232.754948,1194.7868 233.536282,1194.52147 C234.318282,1194.2568 235.086282,1193.94813 235.835615,1193.6008 C236.584948,1193.25347 237.315615,1192.8668 238.026948,1192.44413 C238.382282,1192.2328 238.731615,1192.01213 239.074948,1191.7808 C239.418948,1191.54813 239.754282,1191.3108 240.082282,1191.04547 C240.116282,1191.01813 240.138282,1190.97747 240.140282,1190.9308 C240.144282,1190.84413 240.077615,1190.77147 239.990948,1190.76747 L239.988282,1190.76747 L239.101615,1190.72813 C239.774948,1189.4288 240.294282,1188.0488 240.623615,1186.6208 L240.736282,1186.0508 C240.754282,1185.95547 240.774948,1185.86147 240.790282,1185.76547 L240.831615,1185.47813 C240.856948,1185.28613 240.890282,1185.09413 240.908948,1184.90147 L240.960282,1184.3228 C241.010282,1183.55013 241.017615,1182.7728 240.956948,1182.00013'
-            id='Fill-1'
-          ></path>
-        </g>
-      </g>
-    </svg>
-  </div>
-);
-Rare.propTypes = RarePropTypes;
-
-const VeryRarePropTypes = { color: CorgiType.color };
-
-export const VeryRare = ({ color }) => (
-  <div>
-    <svg
-      width='144px'
-      height='48px'
-      viewBox='0 0 144 48'
-      version='1.1'
-      xmlns='http://www.w3.org/2000/svg'
-      xlink='http://www.w3.org/1999/xlink'
-    >
-      <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-        <g id='result' transform='translate(-121.000000, -1239.000000)' fill={color}>
-          <path
-            d='M139.074948,1284.0508 C139.088948,1284.01813 139.103615,1283.98547 139.118282,1283.9528 L139.120282,1283.9488 C139.135615,1283.91547 139.150948,1283.8808 139.167615,1283.8468 C139.167615,1283.84613 139.167615,1283.84547 139.168282,1283.8448 C139.183615,1283.81013 139.200282,1283.7748 139.217615,1283.73813 C139.344948,1283.4648 139.494948,1283.1608 139.663615,1282.84547 C139.637615,1283.10213 139.588948,1283.35813 139.519615,1283.60813 C139.446948,1283.8708 139.352948,1284.1348 139.241615,1284.38013 C139.140948,1284.3708 139.038948,1284.3668 138.937615,1284.36547 C138.978282,1284.26947 139.024282,1284.1648 139.074948,1284.0508 M264.956948,1250.00013 C264.833615,1248.45547 264.507615,1246.92547 263.969615,1245.47347 C263.703615,1244.74613 263.390282,1244.0368 263.033615,1243.34947 C262.857615,1243.00547 262.666282,1242.66813 262.464948,1242.3368 C262.262282,1242.0048 262.054282,1241.6828 261.813615,1241.36547 C261.783615,1241.3248 261.735615,1241.29813 261.681615,1241.2968 C261.586948,1241.29413 261.506948,1241.36947 261.504948,1241.46413 L261.504948,1241.47147 C261.502948,1241.5288 261.496948,1241.6028 261.489615,1241.6708 C261.483615,1241.73947 261.472948,1241.81013 261.464282,1241.88013 C261.445615,1242.02013 261.420948,1242.1608 261.394948,1242.3008 C261.342282,1242.58147 261.280282,1242.86147 261.212948,1243.1408 C261.143615,1243.41947 261.072282,1243.69813 260.992282,1243.9748 C260.916948,1244.2528 260.829615,1244.52747 260.745615,1244.80347 C260.401615,1245.90413 260.004948,1246.99013 259.586282,1248.0668 L259.266948,1248.87347 L258.935615,1249.67547 C258.720282,1250.21147 258.485615,1250.74013 258.258282,1251.2728 C257.808948,1252.33147 257.242948,1253.3348 256.526948,1254.2348 C255.818282,1255.1388 254.996948,1255.95347 254.102282,1256.67747 C253.648282,1257.0328 253.187615,1257.37947 252.702282,1257.69147 C252.224282,1258.0128 251.726282,1258.30613 251.221615,1258.58613 C250.714948,1258.86347 250.196282,1259.1188 249.671615,1259.36013 C249.144282,1259.59613 248.609615,1259.8168 248.066948,1260.01947 C247.714948,1260.15147 247.358282,1260.27347 247.000282,1260.38947 C246.933615,1260.3288 246.866948,1260.27013 246.800282,1260.21147 C246.767615,1260.18213 246.734948,1260.15213 246.701615,1260.1228 C246.637615,1260.06747 246.572948,1260.01413 246.508282,1259.95947 C246.468948,1259.92613 246.428948,1259.89213 246.388948,1259.85947 C246.326948,1259.80813 246.263615,1259.7588 246.200948,1259.7088 C246.158282,1259.67547 246.115615,1259.64013 246.072282,1259.60613 C246.010282,1259.5588 245.948948,1259.51213 245.886948,1259.46547 C245.841615,1259.43147 245.796948,1259.39813 245.751615,1259.36413 C245.690282,1259.3188 245.628948,1259.27613 245.566948,1259.23147 C245.520282,1259.1988 245.474282,1259.16547 245.426948,1259.1328 C245.366282,1259.09013 245.304282,1259.0488 245.242282,1259.00813 C245.194948,1258.97547 245.146948,1258.9428 245.098948,1258.91147 C245.037615,1258.8708 244.976282,1258.83213 244.914282,1258.7928 C244.865615,1258.76213 244.816948,1258.7308 244.768282,1258.7008 C244.706282,1258.6628 244.644282,1258.62547 244.582948,1258.5888 C244.532948,1258.5588 244.483615,1258.5288 244.433615,1258.49947 C244.371615,1258.46347 244.310282,1258.42813 244.248282,1258.39347 C244.197615,1258.3648 244.146948,1258.33613 244.096282,1258.30813 C244.034282,1258.27413 243.972282,1258.2408 243.909615,1258.2068 C243.858948,1258.18013 243.808282,1258.1528 243.756948,1258.1268 C243.694948,1258.09413 243.632282,1258.0628 243.568948,1258.03147 C243.517615,1258.00547 243.466948,1257.97947 243.415615,1257.95413 C243.352282,1257.92413 243.288948,1257.89413 243.226948,1257.86413 C243.174282,1257.84013 243.122948,1257.8148 243.070948,1257.7908 C243.007615,1257.76213 242.944948,1257.73413 242.881615,1257.70547 C242.829615,1257.6828 242.777615,1257.65947 242.724948,1257.6368 L242.534948,1257.5568 C242.482282,1257.53547 242.429615,1257.5128 242.376948,1257.49147 C242.313615,1257.46613 242.250282,1257.4408 242.186948,1257.41613 C242.134282,1257.39547 242.081615,1257.3748 242.028948,1257.3548 C241.964948,1257.3308 241.900948,1257.30747 241.837615,1257.28413 C241.783615,1257.26413 241.730948,1257.24547 241.677615,1257.2268 C241.614282,1257.20347 241.550282,1257.18147 241.486282,1257.15947 C241.432948,1257.14213 241.380282,1257.12347 241.326948,1257.10613 C241.262948,1257.0848 241.198948,1257.0648 241.134948,1257.04347 C241.081615,1257.0268 241.028282,1257.01013 240.974948,1256.99347 C240.910948,1256.97413 240.846282,1256.95413 240.782282,1256.93547 C240.728948,1256.92013 240.675615,1256.90413 240.622282,1256.88947 C240.558282,1256.8708 240.494282,1256.8528 240.429615,1256.83547 C240.376282,1256.82013 240.322948,1256.80613 240.269615,1256.79213 C240.204948,1256.77547 240.140948,1256.7588 240.076948,1256.74213 C240.023615,1256.7288 239.970282,1256.71547 239.917615,1256.7028 C239.852282,1256.6868 239.787615,1256.67147 239.723615,1256.65613 C239.670948,1256.64413 239.617615,1256.63147 239.564282,1256.62013 L239.370948,1256.5768 C239.318282,1256.56547 239.264948,1256.55547 239.212282,1256.54413 C239.147615,1256.5308 239.083615,1256.51747 239.018948,1256.50547 C238.966282,1256.4948 238.913615,1256.4848 238.860948,1256.47547 C238.796282,1256.4628 238.731615,1256.45147 238.667615,1256.43947 C238.615615,1256.43013 238.562282,1256.42213 238.510282,1256.4128 C238.446282,1256.40147 238.381615,1256.3908 238.316948,1256.3808 C238.264948,1256.37213 238.212948,1256.3648 238.161615,1256.3568 C238.096948,1256.3468 238.032282,1256.33747 237.967615,1256.32813 C237.916282,1256.32013 237.864948,1256.31413 237.812948,1256.3068 C237.748948,1256.29813 237.684282,1256.2888 237.620282,1256.28147 C237.568948,1256.2748 237.518282,1256.2688 237.466948,1256.2628 C237.402282,1256.25547 237.338282,1256.24747 237.273615,1256.24013 C237.224282,1256.23413 237.174282,1256.23013 237.124282,1256.2248 C237.059615,1256.21813 236.994282,1256.2108 236.928948,1256.20413 C236.880948,1256.20013 236.833615,1256.19613 236.784948,1256.19213 C236.718948,1256.18613 236.652948,1256.18013 236.586948,1256.1748 L236.448282,1256.1648 C236.380948,1256.15947 236.313615,1256.15413 236.247615,1256.14947 C236.202282,1256.14747 236.158282,1256.1448 236.114282,1256.14213 C236.046282,1256.13813 235.977615,1256.13347 235.910282,1256.13013 C235.869615,1256.12813 235.830282,1256.1268 235.789615,1256.1248 C235.718282,1256.12147 235.646282,1256.11813 235.575615,1256.11547 C235.548948,1256.1148 235.522282,1256.11413 235.496282,1256.1128 C235.220948,1256.10347 234.948948,1256.09813 234.680282,1256.09747 C234.652948,1256.09747 234.625615,1256.0968 234.598948,1256.0968 C234.531615,1256.0968 234.464948,1256.09747 234.397615,1256.09813 C234.357615,1256.0988 234.316948,1256.0988 234.277615,1256.09947 C234.215615,1256.10013 234.154282,1256.10147 234.092948,1256.1028 C234.048948,1256.10347 234.004282,1256.10413 233.959615,1256.10547 C233.901615,1256.1068 233.844282,1256.1088 233.786948,1256.1108 C233.739615,1256.11213 233.692948,1256.11347 233.645615,1256.1148 C233.588948,1256.11747 233.533615,1256.11947 233.478282,1256.12147 C233.430948,1256.12347 233.384282,1256.12547 233.337615,1256.12747 C233.282282,1256.13013 233.227615,1256.13347 233.172282,1256.13613 C233.125615,1256.1388 233.078948,1256.14147 233.032948,1256.14413 C232.978282,1256.14747 232.925615,1256.15013 232.872282,1256.15347 C232.825615,1256.1568 232.779615,1256.15947 232.732948,1256.1628 L232.576948,1256.1748 C232.530948,1256.17747 232.484282,1256.1808 232.438282,1256.1848 C232.387615,1256.1888 232.337615,1256.19347 232.286948,1256.19747 C232.240948,1256.20147 232.194948,1256.2048 232.148948,1256.2088 C232.099615,1256.21347 232.050948,1256.21813 232.002282,1256.2228 C231.956282,1256.22747 231.910948,1256.23147 231.865615,1256.23547 C231.816948,1256.2408 231.769615,1256.24613 231.721615,1256.2508 C231.676948,1256.25547 231.632282,1256.26013 231.588282,1256.2648 C231.540282,1256.27013 231.494282,1256.27547 231.447615,1256.2808 C231.402948,1256.28613 231.359615,1256.2908 231.314948,1256.29613 C231.270282,1256.30147 231.224948,1256.3068 231.179615,1256.3128 C231.136282,1256.31747 231.092948,1256.3228 231.049615,1256.3288 C231.005615,1256.33413 230.960948,1256.34013 230.917615,1256.34547 C230.874948,1256.35147 230.832948,1256.3568 230.790948,1256.3628 C230.747615,1256.3688 230.704282,1256.3748 230.661615,1256.3808 C230.620282,1256.3868 230.578948,1256.3928 230.538282,1256.39813 C230.496282,1256.40413 230.454948,1256.4108 230.413615,1256.4168 C230.372948,1256.42347 230.332948,1256.4288 230.292948,1256.4348 C230.252282,1256.44147 230.211615,1256.44813 230.171615,1256.45347 C230.132282,1256.46013 230.092948,1256.4668 230.054948,1256.4728 C230.014948,1256.47947 229.976282,1256.48547 229.937615,1256.49213 C229.898948,1256.4988 229.860948,1256.50547 229.822948,1256.5108 C229.784948,1256.51747 229.748282,1256.52413 229.711615,1256.5308 C229.672948,1256.5368 229.635615,1256.54347 229.598282,1256.5508 C229.562282,1256.5568 229.527615,1256.56347 229.492948,1256.56947 C229.455615,1256.57613 229.419615,1256.5828 229.382948,1256.59013 C229.348282,1256.59613 229.313615,1256.6028 229.280282,1256.6088 C229.245615,1256.61613 229.210948,1256.62213 229.176282,1256.6288 L229.076948,1256.64813 C229.042948,1256.6548 229.009615,1256.66213 228.976948,1256.6688 C228.944948,1256.67547 228.913615,1256.6808 228.882282,1256.68747 C228.849615,1256.69413 228.817615,1256.70147 228.785615,1256.70747 C228.755615,1256.71347 228.726282,1256.72013 228.696948,1256.7268 C228.664948,1256.7328 228.634282,1256.73947 228.602948,1256.74613 C228.574948,1256.7528 228.546948,1256.7588 228.518948,1256.76413 C228.488948,1256.77147 228.460282,1256.77747 228.430948,1256.7848 C228.412282,1256.78813 228.392948,1256.79213 228.373615,1256.7968 C227.396948,1257.0168 226.399615,1257.13147 225.398282,1257.13147 L150.477615,1257.13147 C147.060948,1255.87413 144.922948,1257.28147 144.922948,1257.28147 C143.187615,1254.9668 140.982282,1252.1848 138.698282,1249.8248 C138.632282,1249.75747 138.566948,1249.6888 138.501615,1249.62147 C139.114948,1248.22013 138.644282,1246.1108 138.644282,1246.1108 C139.160948,1244.62213 139.616282,1240.18613 137.489615,1239.1528 C135.362948,1238.12013 134.207615,1242.6168 133.582282,1243.80147 C132.956282,1244.9868 132.394282,1244.8048 132.394282,1244.8048 C132.570948,1245.13147 132.496282,1245.45747 132.397615,1245.67947 C131.870282,1245.6868 131.273615,1245.72613 130.614282,1245.74813 C130.067615,1245.76613 129.478282,1245.77213 128.847615,1245.7388 C128.575615,1245.7248 128.296282,1245.70347 128.008282,1245.67213 C127.980282,1245.64613 127.950282,1245.6168 127.916948,1245.58613 C127.852282,1245.52613 127.778282,1245.4568 127.696948,1245.3808 C127.685615,1245.37013 127.672948,1245.35947 127.662282,1245.3488 C127.632948,1245.32147 127.603615,1245.29347 127.572948,1245.26413 C127.557615,1245.24947 127.540282,1245.23347 127.524948,1245.21813 C127.496948,1245.19213 127.469615,1245.16547 127.440948,1245.13747 C127.422282,1245.12013 127.403615,1245.10147 127.385615,1245.08347 C127.357615,1245.05613 127.329615,1245.02813 127.300948,1244.99947 C127.281615,1244.9808 127.262282,1244.9608 127.242948,1244.94147 C127.213615,1244.91147 127.184282,1244.8808 127.154282,1244.85013 L127.097615,1244.79147 C127.064948,1244.75747 127.032282,1244.72213 126.999615,1244.68613 C126.982948,1244.66947 126.967615,1244.65213 126.950948,1244.6348 C126.902282,1244.58147 126.853615,1244.5268 126.805615,1244.4708 C126.803615,1244.4688 126.802282,1244.4668 126.800282,1244.4648 C126.754282,1244.41147 126.708282,1244.3568 126.662948,1244.30147 C126.646948,1244.2828 126.632948,1244.26347 126.617615,1244.2448 C126.586282,1244.20547 126.554948,1244.16613 126.524282,1244.12547 C126.506282,1244.1028 126.489615,1244.08013 126.472282,1244.0568 C126.444948,1244.01947 126.417615,1243.98213 126.390948,1243.9448 C126.372948,1243.92013 126.356282,1243.89547 126.338948,1243.87013 C126.312948,1243.83347 126.288282,1243.79547 126.264282,1243.75813 C126.248282,1243.73213 126.231615,1243.7068 126.216282,1243.68147 C126.192282,1243.64347 126.169615,1243.60413 126.146282,1243.5648 C126.132282,1243.54013 126.117615,1243.51547 126.104282,1243.4908 C126.080948,1243.44813 126.059615,1243.4048 126.037615,1243.36213 C126.026948,1243.3408 126.015615,1243.3188 126.005615,1243.29747 C125.975615,1243.2328 125.946948,1243.16813 125.921615,1243.1028 C125.366282,1241.67547 123.603615,1238.9708 122.175615,1239.57813 C120.747615,1240.18613 121.020948,1243.83213 121.020948,1243.83213 C121.020948,1243.83213 121.222282,1245.07813 121,1245.38147 C121,1245.38147 121.385615,1245.74613 121.537615,1246.50613 C121.669615,1247.16547 121.247615,1249.33947 122.320948,1250.25613 C122.071615,1250.7748 121.936948,1251.12013 121.936948,1251.12013 C121.936948,1251.12013 122.266948,1252.3908 121.172948,1252.7148 C121.172948,1252.7148 122.590948,1252.7148 122.388282,1253.5648 C122.185615,1254.41613 121.780948,1256.2568 121.982948,1257.28147 C121.982948,1257.28147 121.989615,1257.27213 122.000948,1257.25813 C122.012282,1257.24213 122.030282,1257.22013 122.052282,1257.19413 C122.064948,1257.17947 122.078948,1257.16347 122.094282,1257.14747 C122.112282,1257.12747 122.131615,1257.10813 122.152948,1257.08813 C122.164948,1257.07613 122.177615,1257.0648 122.190948,1257.05347 C122.243615,1257.0068 122.303615,1256.96213 122.366282,1256.93147 C122.398948,1256.91547 122.433615,1256.9028 122.468282,1256.89613 C122.487615,1256.89147 122.506948,1256.8888 122.526948,1256.88813 C122.546948,1256.88747 122.566948,1256.89013 122.587615,1256.89413 C122.593615,1256.89547 122.600948,1256.89613 122.607615,1256.89813 C122.630282,1256.90413 122.652948,1256.91547 122.674948,1256.9288 C122.678948,1256.9308 122.682948,1256.9328 122.686282,1256.93547 C122.691615,1256.9388 122.696282,1256.9428 122.700282,1256.9468 C122.708948,1256.9528 122.716948,1256.95813 122.724948,1256.96547 C122.726948,1256.9668 122.727615,1256.9688 122.729615,1256.9708 C122.753615,1256.9928 122.776282,1257.01947 122.798282,1257.0528 C122.816948,1257.0808 122.834948,1257.11413 122.852282,1257.1508 C122.856948,1257.16013 122.861615,1257.1688 122.865615,1257.17813 C122.866948,1257.18213 122.868948,1257.18747 122.870948,1257.19213 C122.900282,1257.2628 122.927615,1257.34747 122.951615,1257.44947 C122.963615,1257.49813 122.971615,1257.54747 122.979615,1257.5968 C122.984948,1257.6328 122.990282,1257.6688 122.994948,1257.70413 C123.014948,1257.8788 123.018282,1258.0548 123.003615,1258.23013 C122.960948,1258.7268 122.885615,1259.88613 122.935615,1261.42413 L122.935615,1261.4248 C122.936282,1261.43147 122.936282,1261.43947 122.936948,1261.44747 C122.938948,1261.52013 122.942282,1261.59347 122.944948,1261.6688 L122.948948,1261.7568 C122.950948,1261.8108 122.952948,1261.8648 122.955615,1261.91947 C122.957615,1261.95347 122.959615,1261.98747 122.960948,1262.02213 C122.964282,1262.07347 122.966948,1262.1248 122.969615,1262.1768 C122.971615,1262.2128 122.974282,1262.24947 122.976282,1262.28547 C122.978948,1262.33747 122.982282,1262.38947 122.985615,1262.44213 C122.988282,1262.4788 122.990948,1262.51547 122.992948,1262.5528 C122.996948,1262.60613 123.000948,1262.66013 123.004948,1262.71413 C123.006948,1262.7508 123.010282,1262.7868 123.012948,1262.82347 C123.017615,1262.8828 123.022948,1262.94347 123.027615,1263.00347 C123.030282,1263.03547 123.032948,1263.06547 123.034948,1263.09747 C123.043615,1263.19013 123.052282,1263.2828 123.060948,1263.37613 C123.060948,1263.3788 123.060948,1263.38213 123.061615,1263.3848 C123.070282,1263.47613 123.079615,1263.56747 123.088948,1263.65947 C123.092282,1263.6908 123.096282,1263.7228 123.099615,1263.75413 C123.106948,1263.81813 123.114282,1263.8828 123.120948,1263.94813 C123.125615,1263.9848 123.130282,1264.0228 123.134948,1264.0608 C123.142282,1264.12013 123.148948,1264.18013 123.156948,1264.2408 C123.161615,1264.2808 123.166948,1264.32013 123.172282,1264.36013 C123.179615,1264.41947 123.187615,1264.47947 123.196282,1264.5388 C123.200948,1264.57947 123.206948,1264.62013 123.212282,1264.66013 C123.220948,1264.72147 123.229615,1264.78347 123.238948,1264.84413 C123.244948,1264.88347 123.250948,1264.9228 123.256282,1264.96213 C123.266948,1265.02947 123.277615,1265.0968 123.288282,1265.1648 C123.293615,1265.1988 123.298948,1265.23147 123.304282,1265.26547 C123.320282,1265.36747 123.338282,1265.46947 123.355615,1265.57213 C123.355615,1265.57213 124.252282,1272.40347 127.106948,1275.1188 C127.106948,1275.1188 126.925615,1274.65613 126.978282,1274.34413 C131.507615,1280.41147 134.722948,1280.68413 134.722948,1280.68413 L136.100948,1283.25147 L136.052282,1283.49547 C135.956948,1283.51213 135.856282,1283.5328 135.751615,1283.5588 C135.490948,1283.62147 135.208282,1283.71547 134.943615,1283.85347 C134.890948,1283.8808 134.838948,1283.91013 134.786948,1283.94147 C134.696282,1283.99747 134.608282,1284.06013 134.524948,1284.12813 C134.426282,1284.2088 134.334948,1284.29813 134.253615,1284.39813 C134.092282,1284.59413 133.968948,1284.82947 133.904282,1285.11213 C133.862948,1285.29147 133.956282,1285.4648 134.109615,1285.54013 C134.160282,1285.56547 134.218282,1285.58013 134.279615,1285.58013 L136.833615,1285.58013 C136.791615,1285.67547 136.755615,1285.77147 136.726948,1285.87013 C136.707615,1285.93547 136.691615,1286.00213 136.677615,1286.0688 L136.668282,1286.11947 C136.665615,1286.13547 136.662282,1286.1528 136.659615,1286.17613 C136.653615,1286.21947 136.654948,1286.26413 136.660282,1286.30813 C136.682282,1286.4828 136.800282,1286.64147 136.962282,1286.71347 C137.002948,1286.73147 137.044948,1286.74547 137.088948,1286.75213 C137.130282,1286.76013 137.180948,1286.75947 137.208282,1286.75947 L137.792282,1286.7588 L139.349615,1286.75413 L142.464948,1286.74547 L142.659615,1286.7448 L142.756948,1286.74413 C142.790948,1286.74413 142.816948,1286.74547 142.864282,1286.7428 C143.032948,1286.73013 143.194282,1286.62747 143.272282,1286.47213 C143.290948,1286.43547 143.306948,1286.3868 143.312948,1286.35747 C143.321615,1286.32413 143.330948,1286.2908 143.337615,1286.25813 C143.352282,1286.1908 143.364282,1286.12413 143.372948,1286.05747 C143.392282,1285.92347 143.400948,1285.7888 143.404282,1285.6548 C143.408282,1285.13947 143.332948,1284.63813 143.204948,1284.1428 C143.716282,1283.62547 144.106282,1283.0028 144.391615,1282.34413 C144.546282,1281.9868 144.672282,1281.6188 144.776282,1281.24413 C145.059615,1282.09213 144.690948,1282.93747 144.690948,1282.93747 C146.686948,1282.6048 148.124282,1281.25347 148.124282,1281.25347 C148.930948,1281.53813 149.314282,1281.53947 149.314282,1281.53947 C149.932282,1281.15947 150.477615,1281.25347 150.477615,1281.25347 L225.338282,1281.25347 C226.170282,1281.25347 225.615615,1281.80813 225.615615,1281.80813 C227.468948,1281.80813 228.906282,1280.2348 228.906282,1280.2348 L228.666948,1280.96947 C231.565615,1279.78147 231.803615,1279.87613 231.803615,1279.87613 C231.751615,1280.12547 231.613615,1280.2668 231.613615,1280.2668 C231.714282,1280.22547 231.811615,1280.18213 231.906948,1280.1388 C231.935615,1280.12547 231.962282,1280.11213 231.990948,1280.0988 C232.056282,1280.06813 232.120948,1280.03813 232.183615,1280.00747 C232.215615,1279.99147 232.245615,1279.97613 232.276948,1279.95947 C232.333615,1279.9308 232.390282,1279.9028 232.444282,1279.87347 C232.476948,1279.8568 232.506282,1279.84013 232.537615,1279.82213 C232.588948,1279.7948 232.640282,1279.7668 232.688948,1279.7388 C232.720282,1279.72147 232.748948,1279.70347 232.779615,1279.68613 C232.826282,1279.65813 232.873615,1279.6308 232.918948,1279.60347 C232.948948,1279.5848 232.977615,1279.56613 233.006948,1279.54747 C233.048948,1279.52147 233.091615,1279.4948 233.132282,1279.46813 C233.162948,1279.4488 233.190948,1279.4288 233.220282,1279.40947 C233.256948,1279.38413 233.295615,1279.35947 233.331615,1279.33413 C233.361615,1279.31347 233.389615,1279.2928 233.418282,1279.27213 C233.450948,1279.2488 233.484282,1279.22547 233.515615,1279.20213 C233.545615,1279.17947 233.572948,1279.15813 233.601615,1279.1368 C233.629615,1279.11547 233.658948,1279.0928 233.686282,1279.07147 C233.715615,1279.0488 233.742948,1279.0268 233.770948,1279.00413 C233.794948,1278.98413 233.819615,1278.9648 233.842948,1278.94547 C233.872282,1278.9208 233.900282,1278.8968 233.928282,1278.87213 C233.947615,1278.85613 233.967615,1278.83947 233.985615,1278.8228 C234.015615,1278.7968 234.043615,1278.77147 234.071615,1278.74547 C234.086282,1278.73213 234.101615,1278.7188 234.116282,1278.70547 C234.144948,1278.67747 234.172948,1278.6508 234.200948,1278.62413 C234.211615,1278.61413 234.222948,1278.60347 234.232948,1278.59347 C234.262282,1278.5648 234.290282,1278.5368 234.317615,1278.5088 C234.324282,1278.50213 234.331615,1278.4948 234.337615,1278.48813 C234.366948,1278.45747 234.394282,1278.42813 234.420948,1278.40013 C234.423615,1278.3968 234.426948,1278.39347 234.430282,1278.39013 C234.436282,1278.38347 234.442948,1278.37613 234.448948,1278.37013 C234.756282,1278.7168 235.084948,1279.04613 235.442282,1279.34613 L235.692282,1279.55213 C235.778282,1279.61813 235.868282,1279.6768 235.956282,1279.74013 C236.044948,1279.8028 236.132948,1279.86613 236.228948,1279.91747 C236.323615,1279.97147 236.415615,1280.03013 236.512282,1280.07947 L236.809615,1280.21947 C236.913615,1280.25747 237.014948,1280.2988 237.120282,1280.33147 C237.228282,1280.36013 237.334282,1280.3928 237.446948,1280.40813 L237.530282,1280.42213 L237.616282,1280.4288 C237.634282,1280.43013 237.652282,1280.43147 237.671615,1280.43213 C237.717615,1280.80947 237.752948,1281.19013 237.778948,1281.56947 C237.806948,1281.9868 237.822948,1282.40547 237.818948,1282.82213 C237.813615,1283.23813 237.791615,1283.6548 237.719615,1284.05747 C237.683615,1284.25813 237.634282,1284.45547 237.556948,1284.6328 C237.484282,1284.79813 237.377615,1284.93813 237.248282,1285.00147 C236.943615,1284.94413 236.644948,1284.92213 236.338948,1284.93213 C236.019615,1284.94547 235.694282,1285.00013 235.388948,1285.12947 C235.236282,1285.1948 235.089615,1285.28013 234.958948,1285.38747 C234.828948,1285.49547 234.716282,1285.62547 234.628282,1285.76947 C234.540948,1285.91413 234.473615,1286.07013 234.433615,1286.2308 C234.427615,1286.25147 234.422282,1286.2688 234.416948,1286.29213 C234.412948,1286.3168 234.406948,1286.34147 234.405615,1286.3668 C234.401615,1286.4168 234.402282,1286.46747 234.412282,1286.5168 C234.429615,1286.61613 234.472282,1286.7108 234.536282,1286.7868 C234.600282,1286.86413 234.683615,1286.9248 234.777615,1286.9608 C234.823615,1286.97947 234.872948,1286.99013 234.922282,1286.99613 C234.946948,1286.99813 234.974948,1287.00013 234.994948,1287.00013 L235.052948,1287.00013 L235.516282,1286.9988 L238.296282,1286.99347 L240.149615,1286.99013 L241.076282,1286.98813 L241.192282,1286.98813 C241.232948,1286.98747 241.262948,1286.98947 241.316282,1286.98613 C241.414282,1286.97947 241.510282,1286.94613 241.590948,1286.89147 C241.672282,1286.83747 241.740282,1286.76147 241.783615,1286.67413 C241.806282,1286.63013 241.822282,1286.58347 241.832282,1286.53613 C241.837615,1286.50947 241.839615,1286.4908 241.842948,1286.47213 L241.851615,1286.41547 L241.978948,1285.58013 L243.523615,1285.58013 C243.557615,1285.58013 243.590282,1285.5788 243.623615,1285.57613 C243.689615,1285.57147 243.755615,1285.56213 243.818948,1285.54747 C243.850948,1285.54013 243.882282,1285.53213 243.913615,1285.5228 C244.037615,1285.4848 244.155615,1285.42947 244.263615,1285.35947 C244.290282,1285.34213 244.316948,1285.3228 244.342282,1285.30347 C244.444948,1285.22547 244.536282,1285.13213 244.613615,1285.0268 C244.652282,1284.97413 244.686948,1284.91813 244.718282,1284.85947 C244.732948,1284.82947 244.748282,1284.79947 244.760948,1284.7688 C244.803615,1284.67147 244.848282,1284.56613 244.894948,1284.45747 C244.949615,1284.33013 245.005615,1284.1968 245.062282,1284.05813 C245.112282,1283.93813 245.162282,1283.81613 245.212282,1283.6908 C245.307615,1283.45347 245.402282,1283.2088 245.492282,1282.96613 C245.967615,1281.68213 246.204948,1279.97147 246.204948,1279.97147 C246.319615,1279.91813 246.431615,1279.86213 246.542948,1279.8048 L246.636948,1279.75613 C246.722282,1279.7108 246.806282,1279.6648 246.890282,1279.61813 C246.917615,1279.60213 246.945615,1279.58747 246.972948,1279.57147 C247.195615,1279.44347 247.412282,1279.30813 247.622282,1279.16547 C247.635615,1279.15613 247.648282,1279.14747 247.662282,1279.13813 C247.759615,1279.0708 247.856282,1279.0028 247.950282,1278.93213 C247.960282,1278.92613 247.970282,1278.9188 247.978948,1278.91147 C251.094948,1276.5948 252.684282,1272.62613 251.729615,1268.29747 C251.727615,1268.2888 251.726948,1268.2808 251.724948,1268.27213 C251.703615,1268.18013 251.681615,1268.08813 251.658948,1267.99547 C251.654282,1267.97613 251.649615,1267.95613 251.644282,1267.9368 C251.623615,1267.85147 251.600282,1267.76613 251.576948,1267.6808 C251.569615,1267.65413 251.562282,1267.6268 251.554948,1267.60013 C251.532948,1267.52013 251.508282,1267.44013 251.484282,1267.36013 C251.474948,1267.32747 251.464948,1267.29413 251.454282,1267.26147 C251.430948,1267.18547 251.405615,1267.10947 251.380948,1267.03347 C251.368282,1266.9968 251.356282,1266.95947 251.343615,1266.9228 C251.318282,1266.84947 251.292282,1266.7768 251.266282,1266.70347 C251.252282,1266.66347 251.236948,1266.62213 251.222282,1266.58213 C251.196282,1266.51147 251.168948,1266.44147 251.141615,1266.3708 C251.124282,1266.32747 251.106948,1266.2848 251.089615,1266.2408 C251.061615,1266.1728 251.033615,1266.10413 251.005615,1266.03613 L250.946948,1265.8988 C250.918282,1265.8328 250.888282,1265.76613 250.858282,1265.69947 C250.836948,1265.65147 250.814282,1265.60413 250.792282,1265.55747 C250.762282,1265.49213 250.731615,1265.4268 250.700282,1265.36213 C250.676282,1265.3128 250.651615,1265.26347 250.627615,1265.21413 C250.595615,1265.15013 250.563615,1265.0868 250.531615,1265.02347 C250.505615,1264.9728 250.477615,1264.92213 250.450948,1264.8708 C250.418282,1264.8088 250.384948,1264.74613 250.350948,1264.68347 C250.322282,1264.63213 250.292948,1264.58013 250.263615,1264.52813 C250.228948,1264.4668 250.194948,1264.40547 250.159615,1264.34413 C250.128282,1264.2908 250.096282,1264.23747 250.064282,1264.18413 C250.028948,1264.12413 249.992948,1264.0648 249.956282,1264.0048 C249.922282,1263.95013 249.888282,1263.89547 249.853615,1263.8408 C249.816282,1263.78213 249.780282,1263.72347 249.741615,1263.66413 C249.705615,1263.60813 249.668282,1263.5528 249.630948,1263.4968 L249.626948,1263.4908 L249.919615,1263.5228 C250.022282,1263.53213 250.124948,1263.53613 250.227615,1263.54213 C251.872282,1263.6528 253.534948,1263.52547 255.148948,1263.17947 C255.958282,1263.0128 256.754948,1262.7868 257.536282,1262.52147 C258.318282,1262.2568 259.086282,1261.94813 259.835615,1261.6008 C260.584948,1261.25347 261.315615,1260.8668 262.026948,1260.44413 C262.382282,1260.2328 262.731615,1260.01213 263.074948,1259.7808 C263.418948,1259.54813 263.754282,1259.3108 264.082282,1259.04547 C264.116282,1259.01813 264.138282,1258.97747 264.140282,1258.9308 C264.144282,1258.84413 264.077615,1258.77147 263.990948,1258.76747 L263.988282,1258.76747 L263.101615,1258.72813 C263.774948,1257.4288 264.294282,1256.0488 264.623615,1254.6208 L264.736282,1254.0508 C264.754282,1253.95547 264.774948,1253.86147 264.790282,1253.76547 L264.831615,1253.47813 C264.856948,1253.28613 264.890282,1253.09413 264.908948,1252.90147 L264.960282,1252.3228 C265.010282,1251.55013 265.017615,1250.7728 264.956948,1250.00013'
-            id='Fill-1'
-          ></path>
-        </g>
-      </g>
-    </svg>
-  </div>
-);
-VeryRare.propTypes = VeryRarePropTypes;
