+++ src/component/utils/corgiPhotos/CorgiRate.jsx
