+++ src/component/utils/corgiPhotos/Uncommon.jsx
